% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{add_px_size}
\alias{add_px_size}
\title{Attach Pixel Size Attribute}
\usage{
add_px_size(object, px.size)
}
\arguments{
\item{object}{any object}

\item{px.size}{a list, a character string, or NULL that is passed to \code{px_size_calc()}.}
}
\value{
The original object with a "px.size" attribute
}
\description{
Attach "px.size" attribute to an object.
}
