% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R, R/utils.R
\name{add_quote}
\alias{add_quote}
\title{Add quotes around a string}
\usage{
add_quote(x)

add_quote(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a character string

an atomic character vector
}
\description{
Add single quotes around a string

Add quotes around a character string
}
