% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weirdo_distributions.R
\name{adjust_prop}
\alias{adjust_prop}
\title{Adjust Proportion Data}
\usage{
adjust_prop(
  x,
  nudge.method = c("replace", "smithson", "add", "subtract", "drop"),
  nudge.size = c("macmillan", "warton_min", "warton_max"),
  trans = c("identity", "logit", "empirical_logit", "probit", "log"),
  bounds = c("both", "zero", "one"),
  na.action = c("ignore", "remove", "as.is", "fail")
)
}
\arguments{
\item{x}{a vector of numeric data bounded between 0 and 1 inclusive.}

\item{nudge.method}{a character value indicating the method of getting rid of 0's and 1's. Valid options are "replace" (default), "smithson", "add", "subtract", and "drop". see details.}

\item{nudge.size}{a numeric value of the size of a small nudge (\eqn{epsilon}). Valid character values can also be supplied to choose the method for estimating \eqn(epsilon). Valid methods are "macmillan" (default), "warton_min", and "warton_max". see details.}

\item{trans}{a character value indicating the transformation to be applied to the nudged data. Valid options are "identity" (default-- no transformation), "logit", "empirical_logit", "probit", and "log". See details.}

\item{bounds}{a character value indicating bounds for which the \code{nudge.method} is applied to. Valid options are "both" (default), "zero", and "one". Only relevant for \code{nudge.method} set as "replace" or "drop".}

\item{na.action}{a character value indicating how to handle missing values. Valid options are "ignore", "remove", "as.is", and "fail". See details.}
}
\value{
a vector of numeric value with the same length as \code{x}
}
\description{
Adjusts proportion data to get rid of 0's and 1's, and optionally transforms the resulting adjusted proportion.
}
\references{
Kubinec, R. 2022. Ordered Beta Regression: A Parsimonious, Well-Fitting Model for Continuous Data with Lower and Upper Bounds. Political Analysis:1–18.
Macmillan, N. A., and C. D. Creelman. 2004. Detection Theory: A User’s Guide. Second edition. Psychology Press, New York.
Smithson, M., and J. Verkuilen. 2006. A better lemon squeezer? Maximum-likelihood regression with beta-distributed dependent variables. Psychological Methods 11:54–71.
Warton, D. I., and F. K. C. Hui. 2011. The arcsine is asinine: the analysis of proportions in ecology. Ecology 92:3–10.
}
