% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model_ML_support.R
\name{dallo}
\alias{dallo}
\alias{rallo}
\alias{pallo}
\alias{qallo}
\title{Neutral 'Bite Size' Distribution Based On Allometric Scaling Laws}
\usage{
dallo(x, min.phi = 0.005, max.phi = 1, a = 14/9, log = FALSE)

rallo(n, min.phi = 0.005, max.phi = 1, a = 14/9)

pallo(
  q,
  min.phi = 0.005,
  max.phi = 1,
  a = 14/9,
  lower.tail = TRUE,
  log.p = FALSE
)

qallo(
  p,
  min.phi = 0.005,
  max.phi = 1,
  a = 14/9,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{x, q}{a vector of quantities}

\item{min.phi}{the minimum bite size in terms of proportion leaf herbivory. Defaults to 0.005 (0.5\%).}

\item{max.phi}{the maximum bite size in terms of proportion leaf herbivory Defaults to 1 (100\%).}

\item{a}{the combined allometric scaling coefficient. Defaults to 14/9.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities p are given as \eqn{\log(p)}.}

\item{n}{number of observations to generate.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P(X \leq x)} otherwise, \eqn{P(X > x)}}

\item{p}{a vector of probabilities}
}
\value{
a vector of numeric values
}
\description{
Density, distribution function, quantile function, and random generation for the neutral 'bite size' distribution based on allometric scaling laws. This distribution is a reparameterization of the truncated Pareto distribution.
}
\details{
The neutral 'bite size' distribution has density function
\deqn{P(\phi) = \frac{1-\alpha}{\phi_M^{1-\alpha} - \phi_m^{1-\alpha}} \phi^{-\alpha}} when \eqn{\alpha \neq 1} and
\deqn{P(\phi) = \frac{1}{\ln{\phi_M} - \ln{\phi_m}} \phi^{-1}} when \eqn{\alpha = 1}
where \eqn{\phi_m} is the minimum 'bite size' and \eqn{\phi_M} is the maximum 'bite size' in terms of proportion leaf herbivory, and \eqn{\alpha} is the combined allometric scaling coefficient defined as \deqn{\alpha = - \frac{\alpha_N + \alpha_S + 1 - \alpha_I}{\alpha_I}.} \eqn{\alpha_N} is the allometric scaling exponent between population density \eqn{N} and body mass \eqn{W} such that \eqn{N \propto W^{\alpha_N}}. A priori value is \eqn{-\frac{3}{4}} according to Damuth’s rule (Damuth 1981). \eqn{\alpha_S} is the allometric scaling exponent of species richness \eqn{S_j} among a body mass class \eqn{W_j} such that \eqn{S_j \propto W_{j}^{\alpha_S}}. A priori value is \eqn{-\frac{2}{3}} according to Hutchinson & MacArthur (1959) and May (1978).\eqn{\alpha_I} is the allometric scaling exponent of whole body metabolic rate \eqn{I_j} among a body mass class \eqn{W_j} such that \eqn{I_j \propto W_{j}^{\alpha_I}}. A priori value is \eqn{\frac{3}{4}} according to Kleiber's law (1932).

The cumulative distribution function is
\deqn{P(\phi \leq q) = \frac{q^{1-\alpha} - \phi_m^{1-\alpha}}{\phi_M^{1-\alpha} - \phi_m^{1-\alpha}}} when \eqn{\alpha \neq 1} and
\deqn{P(\phi \leq q) = \frac{\ln{q} - \ln{\phi_m}}{\ln{\phi_M} - \ln{\phi_m}}} when \eqn{\alpha = 1}.

The neutral 'bite size' distribution is equivalent to a reparameterized truncated Pareto distribution with \eqn{\alpha = a - 1}, \eqn{\phi_m = b}, and \eqn{\phi_M = B}. The truncated Pareto distribution has the density function:
\deqn{P(x) = \frac{ab^a x^{-a-1}}{1-(\frac{b}{B})^a}}
}
\references{
Damuth, J. 1981. Population density and body size in mammals. Nature 290:699–700.

Hutchinson, G. E., and R. H. MacArthur. 1959. A Theoretical Ecological Model of Size Distributions Among Species of Animals. The American Naturalist 93:117–125.

Kleiber, M. 1932. Body size and metabolism. Hilgardia 6:315–353.

May, R. M. 1978. The dynamics and diversity of insect faunas. Page in L. A. Mound and N. Waloff, editors. Diversity of insect faunas. Published for the Royal Entomological Society by Blackwell Scientific Publications; Distributed by Halsted Press, Oxford England; New York.
}
