% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{analyze_holes}
\alias{analyze_holes}
\title{Parse Holes And Extract Within Leaf Herbivory Data}
\usage{
analyze_holes(
  object,
  min_prop = 10^-4,
  plot = FALSE,
  tolerance = 0.1,
  silent = FALSE,
  image_id = NA
)
}
\arguments{
\item{object}{A cimg, pixset, matrix, or array that has been cropped via \code{crop_leaf()}}

\item{min_prop}{a numeric value between 0 and 1 indicating the minimum proportion herbivory threshold for inclusion. This gets rid of artifacts in the image that are too small to actually be real herbivory or biologically non-trivial herbivory. Default is 0.0001.}

\item{plot}{if \code{TRUE} (default is \code{FALSE}), each hole splitted from the image will be plotted side-by-side. If more than 20 holes are detected, the user will be prompted to decide whether to proceed with plotting. Can be slow if the number of holes and image resolution is high.}

\item{tolerance}{a numeric value between 0 and 1 determining if two pixels belong to the same region. Defaults to 0.1.}

\item{silent}{if \code{TRUE} (default is \code{FALSE}), suppress messages.}

\item{image_id}{a character string supplied to name the returned object for book keeping.}
}
\value{
an object of class 'split_herb' and 'list'.

\code{n_holes}: The number of holes. Holes with proportion area less than \code{min_prop} are excluded.

\code{hole_prop}: The proportion area of each hole. Returns NA if \code{n_hole} is 0.

\code{leaf_area}: The leaf area in pixels and mm2.

\code{is.margin}: A vector of logical values indicating whether each parsed hole is marginal or internal. Returns NA if \code{n_hole} is 0.

\code{hole_perimeter}: A \code{n_hole} X 2 matrix of values of the perimeter of each parsed hole returned by the function \code{hole_perimeter()}. The \code{non_leaf_border} argument is set to \code{TRUE}. Returns NA if \code{n_hole} is 0.

\code{hole_centroid}: A \code{n_hole} X 2 matrix of the the centroid coordinates of each parsed hole. Returns NA if \code{n_hole} is 0.

\code{px}: A binary-ized pixset of the supplied \code{object} whose value is above 0.8.

\code{min_prop}: The supplied \code{min_prop} argument

\code{px.size}: The length in millimeter of a pixel. If unsuccessfully extracted from the \code{object} \code{px.size} attribute, only results in pixels will be returned for other function outputs.

\code{imlist}: A list of images of class \code{imlist} of each parsed hole.

\code{image_id}: The supplied \code{image_id} argument
}
\description{
Split pixels into different herbivory regions ("holes") and measure hole and leaf level data. A pre-processing step for use with subsequent functions such as \code{split2ppp()}, \code{nn_dist()}, \code{pair_dist()}, and
}
