% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{array2Image}
\alias{array2Image}
\title{Convert Array to Class 'Image'}
\usage{
array2Image(x)
}
\arguments{
\item{x}{A four dimensional array (as in the format in the package \code{imager}), a cimg, or pixset object.}
}
\value{
an 'Image' object
}
\description{
Convert array formatted as cimg or pixset to an object of class 'Image' from the package \code{EBImage}.
}
