% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{as.cimg_color}
\alias{as.cimg_color}
\title{Convert pixset into a colored cimg}
\usage{
as.cimg_color(img, n = 3)
}
\arguments{
\item{img}{the pixset}

\item{n}{number of color spectra, defaults to 3.}
}
\value{
a colored cimg object
}
\description{
Convert pixset into a colored cimg
}
