% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_named_vector}
\alias{as_named_vector}
\title{Turn data frame to named vector}
\usage{
as_named_vector(x, id_col = 1, val_col = 2)
}
\arguments{
\item{x}{the data.frame}

\item{id_col}{the index or column name (as a character string) of the names}

\item{val_col}{the index or column name (as a character string) of the values}
}
\value{
a named vector
}
\description{
Turn data frame to named vector
}
