% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beta_conjugate}
\alias{beta_conjugate}
\title{Generate uncertainty intervals for binary data using conjugate prior}
\usage{
beta_conjugate(alpha, beta, interval = 0.95, prior = c(1, 1))
}
\arguments{
\item{alpha}{The number of successes}

\item{beta}{The number of failure}

\item{interval}{A numeric value indicating the intervals for upper and lower bounds (default to 0.95).}

\item{prior}{A numeric vector of the alpha and beta value of the conjugate prior. Default is the Bayes-Laplace prior (Tuyl et al. 2008) \eqn{\alpha = 1, \beta = 1}. Other popular choices include 'neutral' prior (Kerman 2011) \eqn{\alpha = 1/3, \beta = 1/3}, and the Jeffreys prior \eqn{\alpha = 0.5, \beta = 0.5}.}
}
\value{
a numeric vector
}
\description{
Generate uncertainty intervals for binary data using conjugate prior (beta distribution)
}
\references{
Kerman, J. 2011. Neutral noninformative and informative conjugate beta and gamma prior distributions. Electronic Journal of Statistics 5:1450–1470.
Tuyl, F., R. Gerlach, and K. Mengersen. 2008. A Comparison of Bayes-Laplace, Jeffreys, and Other Priors: The Case of Zero Events. The American Statistician 62:40–44.
}
