% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bin}
\alias{bin}
\title{Bin values}
\usage{
bin(x, by = 1, value = TRUE, range = c(min(x), max(x)))
}
\arguments{
\item{x}{a vector of numeric values}

\item{by}{the spacing between the bins}

\item{value}{If \code{TRUE}, return the bin value, otherwise return the bin id}

\item{range}{A numeric vector specifying the lower and upper limit of the bins}
}
\value{
a numeric vector
}
\description{
Find the closest lower bin of a vector
}
