% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_vec}
\alias{bind_vec}
\title{Bind a list of vector into data.frame}
\usage{
bind_vec(x, margin = 1L, keep_row_names = TRUE, row_names_as_col = FALSE)
}
\arguments{
\item{x}{a list of vectors}

\item{margin}{the direction for which the vectors will be bound. 1 indicates rows (default) and 2 indicates columns.}

\item{keep_row_names}{if \code{TRUE} (default), the row names will be kept. Otherwise, they will be set to \code{NULL}.}

\item{row_names_as_col}{if a character string is provided, the row name will be added to the data.frame as the first column with that name. If \code{TRUE}, "rownames" will be used as the column name. Otherwise, the no column is added (default is \code{FALSE}).}
}
\value{
a data.frame
}
\description{
Bind a list of vector into a data.frame and keep track of list names and vector names.
}
