% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weirdo_distributions.R
\name{dcb}
\alias{dcb}
\alias{pcb}
\alias{qcb}
\alias{rcb}
\title{Continuous Bernoulli Distribution}
\usage{
dcb(x, lambda, log = FALSE)

pcb(q, lambda, lower.tail = TRUE, log.p = FALSE)

qcb(p, lambda, lower.tail = TRUE, log.p = FALSE)

rcb(n, lambda)
}
\arguments{
\item{x, q}{a vector of quantities defined for \eqn{0 < X < 1}}

\item{lambda}{the shape parameter \eqn{0 < \lambda < 1}}

\item{log, log.p}{a logical value indicating whether to log transform the probabilities. Default is \code{FALSE}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are \eqn{P(X \leq x)} otherwise, \eqn{P(X > x)}}

\item{p}{a vector of probabilities}

\item{n}{number of observations to generate}
}
\value{
a vector of numeric values
}
\description{
Density, distribution function, quantile function, and random generation for a one parameter continuous Bernoulli distribution
}
\details{
The continuous Bernoulli distribution is a one parameter distribution (\eqn{\lambda \in (0,1)}) that has support \eqn{x \in (0,1)}.

The density function is defined as:
\deqn{P(x) = C(\lambda) \lambda^x (1 - \lambda)^{1-x},}

where \eqn{C(\lambda)} is defined as
\deqn{C(\lambda) = 2}
if \eqn{\lambda = \frac{1}{2}}, otherwise,
\deqn{C(\lambda) = \frac{2 \tanh^{-1}(1-2\lambda)}{1-2\lambda}.}

The cumulative distribution function is defined as
\deqn{P(X \leq x) = x}
if \eqn{\lambda = \frac{1}{2}}, otherwise,
\deqn{P(X \leq x) = \frac{\lambda^x (1-\lambda)^{1-x} + \lambda - 1}{2 \lambda - 1}.}
}
\references{
Loaiza-Ganem, G., and J. P. Cunningham. 2019. The continuous Bernoulli: fixing a pervasive error in variational autoencoders. Page Advances in Neural Information Processing Systems. Curran Associates, Inc.
}
