% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{cd}
\alias{cd}
\title{Calculate Coefficient of Dispersion (CD)}
\usage{
cd(x, robust = F, na.rm = F)
}
\arguments{
\item{x}{a vector of numeric data}

\item{robust}{if \code{TRUE}, a more robust for of CD is calculated. Default is \code{FALSE}}

\item{na.rm}{a logical value indicating whether to drop \code{NA} values}
}
\value{
A numeric value
}
\description{
calculate CD for a given mean and variance or vector of data
}
\details{
The coefficient of dispersion \eqn{CD}, a.k.a. variance mean ratio, is calculated as
\deqn{CD = \frac{\sigma^2}{\mu}}
The robust version is calculated as
\deqn{CD = \frac{1}{n} \frac{\sum_i^n |m - x_i|}{m}}
where \eqn{m} is the median of the data
}
