% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{clean_name}
\alias{clean_name}
\title{Remove punctuation and change all characters to lower case}
\usage{
clean_name(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector
}
\description{
Remove punctuation and change all characters to lower case in supplied character vector
}
