% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model_ML_support.R
\name{coef.allo_herb_fit}
\alias{coef.allo_herb_fit}
\title{Extract Model Coefficients}
\usage{
\method{coef}{allo_herb_fit}(object, ..., backtransform = TRUE, se = FALSE)
}
\arguments{
\item{object}{An 'allo_herb_fit' object}

\item{...}{additional arguments}

\item{backtransform}{A logic value indicating whether to back transform coefficients from the scale the coefficient was estimated at to the scale the coefficient is parameterized. Default is TRUE.}

\item{se}{If \code{TRUE}, the standard error will be returned as well. If \code{backtransform = TRUE}, the standard error on the back transformed scale will be approximated using the first order second moment method. See \code{?FOSM}}
}
\value{
A named vector of estimates. If \code{se = TRUE}, a named list of vectors will be returned instead.
}
\description{
extract model coefficients from allo_herb_fit objects
}
