% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_ML_support.R
\name{coef.bite_size_fit}
\alias{coef.bite_size_fit}
\title{Extract Model Coefficients}
\usage{
\method{coef}{bite_size_fit}(object, ..., backtransform = TRUE, se = FALSE)
}
\arguments{
\item{object}{An 'allo_herb_fit' object}

\item{...}{additional arguments}

\item{backtransform}{A logic value indicating whether to back transform coefficients from the scale the coefficient was estimated at to the scale the coefficient is parameterized as for the neutral model. Default is TRUE.}
}
\value{
A list of each fitted family of distribution is returned. For each element of the list, a named vector of estimates is returned. If \code{se = TRUE}, a named list of vectors will be returned instead.
}
\description{
extract model coefficients from bite_size_fit objects
}
