% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{color_index}
\alias{color_index}
\title{Image Color Index From pliman}
\usage{
color_index(img, index = "all", plot = TRUE)
}
\arguments{
\item{img}{a cimg object with three color spectra}

\item{index}{a character string indicating the indices to use. If "all", all indices will be chosen}

\item{plot}{if \code{TRUE} (default), plot the transformed images}
}
\value{
an imlist object
}
\description{
Return and display different transformed colored image, useful for thresholding. Indices are imported from the pliman package. Custom color index can be calculated using \code{imager::imeval()}.
}
\details{
This is an experimental function for internal use at the moment. Credit is due to the package Pliman.
}
