% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{compare_dist}
\alias{compare_dist}
\title{Compare whether two samples are drawn from the same distribution}
\usage{
compare_dist(
  data.list = NULL,
  obs.index = 1,
  pred.index = 2,
  test = c("ks", "kl", "ad", "chisq"),
  digits = 2,
  bin_size = 0.1
)
}
\arguments{
\item{data.list}{A list of two numeric vectors that are compared with each other.}

\item{obs.index}{The index of the numeric vector in the list that correspond to the observations (defaults to 1).}

\item{pred.index}{The index of the numeric vector in the list that correspond to the predictions (defaults to 2).}

\item{test}{A vector of character string indicating which method to use to compare the two samples. "ks" performs the KS test using \code{stats::ks.test()}. "ad" performs the AD test using \code{kSamples::ad.test()}. "kl" computes the KL divergence using \code{philentropy::KL()}. The unit of the KL divergence is "log" by default. "chisq" performs the Chi-square test using \code{stats::chisq.test()}.}

\item{digits}{An integer indicating the number of digits the samples should be rounded to before doing the calculations. This can be important when rounding gets rid of some sampling artifacts in the data.}

\item{bin_size}{The bin size used to calculate KL divergence and Chi-square test.}
}
\value{
A list of named numeric vectors.
}
\description{
Compare whether two samples are drawn from the same distribution, via a two-sample Kolmogorov-Smirnov tests (KS test), Anderson-Darling test (AD test), Chi-square test, and Kullback-Leibler Divergence (KL divergence).
}
