% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neutral_model_ML_support.R
\name{convolve.dist}
\alias{convolve.dist}
\title{Perform k Convolutions of Fourier Transformed Vector}
\usage{
convolve.dist(k, fft.vec, parallel = FALSE)
}
\arguments{
\item{k}{a vector of integers indicating the number of convolutions to apply to a Fourier transformed vector}

\item{fft.vec}{a Fourier transformed vector on the frequency domain}

\item{parallel}{if \code{TRUE} (default is \code{FALSE}), convolve \code{fft.vec} in parallel. Automatically turned off if \code{length(ff.vec)} is less than 50000, as no computational efficiency would be gained.}
}
\value{
A \code{length(fft.vec)} x \code{length(k)} matrix of conditional probabilities
}
\description{
Internal function of \code{dalloT.cond.k.fft.conv()} used to perform \eqn{k} convolutions of a supplied Fourier transformed vector. Parallel computing is supported and requires an already registered cluster.
}
\details{
\deqn{P(\phi_T|k) = \overbrace{P(\phi)*P(\phi)...*P(\phi)}^{\text{k times}} = \mathcal{F}^{-1} [\mathcal{F}[P(\phi)]^k]}
}
