% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{crop_leaf}
\alias{crop_leaf}
\title{Crop Out Non-Leaf Pixels}
\usage{
crop_leaf(object, as.mat = FALSE, invalid = NA, rgb.index = 3L, px.size = NA)
}
\arguments{
\item{object}{a cimg, pixset, or matrix, or array to be cropped.}

\item{as.mat}{if \code{TRUE}, the function returns a matrix of the pixel values of the \code{rgb.index} spectrum. If \code{FALSE} (default), the function returns a grayscale 'cimg' object.}

\item{invalid}{the value to replace non-leaf pixels as. Default is \code{NA}.}

\item{rgb.index}{an integer value \code{1L}, \code{2L}, or \code{3L}, indicating red, green, or blue colors respectively to be identified as non-leaf pixels.}

\item{px.size}{value passed to \code{px_size_calc()}. When set to \code{NA} (default), the 'px.size' attribute is extracted from the supplied \code{object}.}
}
\value{
a cropped 'cimg' object
}
\description{
Crop out non-leaf pixels marked by the color blue as \code{NA}s. Non-leaf pixels need to be cropped out for further analyses.
}
