% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{cut_kmeans}
\alias{cut_kmeans}
\title{Find Threshold For k Clusters Using k-Means}
\usage{
cut_kmeans(x, km = c(1L, 2L))
}
\arguments{
\item{x}{a numeric vector. \code{NA}s are automatically removed.}

\item{km}{an integer vector of length two. The first value selects the the boundary rank in increasing means identified by kmeans clustering (defaults to 1, selecting the boundary that separates the two clusters with the lowest mean). The second value selects the k number of clusters to form (defaults to 2).}
}
\value{
an atomic numeric value of boundary threshold
}
\description{
find threshold for k clusters using k-means
}
\note{
The code for the function is adapted from Simon Barthelme's \code{imager:::cut.kmeans()}. version 0.42.13.
}
