% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{cv}
\alias{cv}
\title{Calculate Coefficient of Variation (CV)}
\usage{
cv(x, method = c(1, 2, 3, 4), na.rm = FALSE)
}
\arguments{
\item{x}{A vector of numeric values}

\item{method}{A value indicating the type of CV estimator}

\item{na.rm}{A logical value indicating whether to ignore \code{NA} values in the \code{x} vector}
}
\value{
A numeric value
}
\description{
calculate the CV for a vector of data
}
\details{
The method option selects one of four CV estimators investigated by Yang et al. (2020).

\strong{CV1}: This is the conventional CV estimator:
\deqn{CV_1 = \frac{\sigma}{\mu}}

\strong{CV2}: This assumes the sample is normally distributed and corrects for the bias using the sample size as in Sokal and Rohlf (1995):
\deqn{CV_2 = CV_1 + \frac{CV_1}{4N}}

\strong{CV3}: This implements Breunig's (2001) method:
\deqn{CV_3 = CV_1 \sqrt{1- \frac{CV_1}{N}(3CV_1 -2\gamma_1)}}
\eqn{\gamma_1} is Pearson's measure of skewness

\strong{CV4}: This implements Bao's (2009) method:
\deqn{CV_4 = CV_1 + \frac{CV_1^3}{N}+ \frac{CV_1}{4N}+ \frac{CV_1^2\gamma_1}{2N} + \frac{CV_1\gamma_2}{8N}}
\eqn{\gamma_2} is Pearson's measure of kurtosis
}
