% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neutral_model_ML_support.R
\name{dalloT.cond.k.gauss.approx}
\alias{dalloT.cond.k.gauss.approx}
\title{Calculate the Conditional Probability Matrix of P(phi.T | k) Via Gaussian Approximation}
\usage{
dalloT.cond.k.gauss.approx(
  phi.T,
  k,
  min.phi = 0.005,
  max.phi = 1,
  a = 14/9,
  log = FALSE
)
}
\arguments{
\item{phi.T}{a vector of cumulative proportion herbivory}

\item{k}{a vector of integer values indicating the number of convolutions to perform}

\item{min.phi}{the minimum bite size in terms of proportion leaf herbivory. Defaults to 0.005 (0.5\%).}

\item{max.phi}{the maximum bite size in terms of proportion leaf herbivory Defaults to 1 (100\%).}

\item{a}{the combined allometry scaling coefficient. Defaults to 14/9.}

\item{log}{if \code{TRUE} (default is \code{FASLE}), return probabilities on the log scale.}
}
\value{
a length(phi.T) X length(k) matrix of probabilities.
}
\description{
Internal function of \code{dalloT.cond.k.fft.conv()} used to calculate \eqn{P(\phi_T | k)} via Gaussian approximation for very large values of \eqn{k.}
}
\details{
Owing to CLT,
\deqn{P(\phi_T | k) \approx \mathcal{N}(\mu = k\overline\phi,\sigma=\sqrt{k\mathbb{Var}[\phi]}),} where
\deqn{\overline{\phi} = \frac{1-\alpha}{2-\alpha} \frac{\phi_M^{2-\alpha} - \phi_m^{2-\alpha}}{\phi_M^{1-\alpha} - \phi_m^{1-\alpha}},} and
\deqn{\mathbb{Var}[\phi]=\frac{\phi_M^{3-\alpha}-\phi_m^{3-\alpha}}{\phi_M^{1-\alpha}-\phi_m^{1-\alpha}}\frac{1-\alpha}{3-\alpha}-(\frac{\phi_M^{2-\alpha}-\phi_m^{2-\alpha}}{\phi_M^{1-\alpha}-\phi_m^{1-\alpha}}\frac{1-\alpha}{2-\alpha})^2.}
}
