% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_averaging_simulation.R
\name{derivative}
\alias{derivative}
\alias{derivative2}
\title{Find derivative}
\usage{
derivative(x, y)

derivative2(x, y)
}
\arguments{
\item{x, y}{a vector of numeric values}
}
\value{
a numeric vector
}
\description{
Find the first or second derivative of the supplied \code{x} and \code{y} vectors using finite difference method.
}
