% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_management.R
\name{detach.herbivar}
\alias{detach.herbivar}
\title{Detach the 'herbivar' package}
\usage{
detach.herbivar(x)
}
\arguments{
\item{x}{placeholder}
}
\value{
\code{NULL}
}
\description{
Unload the 'herbivar' package
}
