% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model_ML_support.R
\name{dk.cond.lambda}
\alias{dk.cond.lambda}
\title{Numerically Estimate Conditional Probability Matrix of P(k | lambda)}
\usage{
dk.cond.lambda(
  k,
  mean.phi.T = NULL,
  min.phi = NULL,
  max.phi = NULL,
  a = NULL,
  k.max.tolerance = 1e-05,
  lambda = NULL,
  log = FALSE
)
}
\arguments{
\item{k}{a vector of integers indicating the number of convolutions to apply to a Fourier transformed vector}

\item{mean.phi.T}{The mean herbivory of the distribution when herbivores are not plant limited.}

\item{min.phi}{the minimum bite size in terms of proportion leaf herbivory. Defaults to 0.005 (0.5\%).}

\item{max.phi}{the maximum bite size in terms of proportion leaf herbivory Defaults to 1 (100\%).}

\item{a}{the combined allometric scaling coefficient. Defaults to 14/9.}

\item{k.max.tolerance}{the tolerance threshold of maximum convolution cut off (ideally probabilities above \code{k.max} convolutions is vanishingly small). See details of \code{?dalloT()}}

\item{lambda}{if \code{NULL} (default), lambda would be estimated from the parameters given. Supplying \code{lambda} is useful to avoid repeated computation when the function is used repeatedly.}

\item{log}{if \code{TRUE}, returns probabilities on the log scale.}
}
\value{
a \code{(length(phi.T))} X \code{length(k)} matrix of conditional probabilities.
}
\description{
Internal function of \code{dalloT()} used to estimate the conditional probability matrix  \eqn{P(k|\lambda)}. Basically a wrapper function for \code{dpois()}.
}
