% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model_ML_support.R
\name{.dalloT.cond.k.fft.conv}
\alias{.dalloT.cond.k.fft.conv}
\title{Numerically Estimate Conditional Probability Matrix of P(phi.T | k)}
\usage{
.dalloT.cond.k.fft.conv(
  phi.T,
  k,
  fft.vec,
  phi,
  log = FALSE,
  min.phi = 0.005,
  max.phi = 1,
  a = 14/9,
  k.fft.limit = 100,
  parallel = FALSE,
  phi.T.index = NULL
)
}
\arguments{
\item{phi.T}{a vector of cumulative proportion herbivory}

\item{k}{a vector of integers indicating the number of convolutions to apply to a Fourier transformed vector}

\item{fft.vec}{a Fourier transformed vector on the frequency domain, specifically that of the probability density of phi.}

\item{phi}{a numeric vector of the proportion herbivory grid used to generate the \code{fft.vec} with \code{dallo()}. Supplied to find closest matched indices for \code{phi.T} for approximate conditional probabilities of \eqn{P(\phi_T | k)}.}

\item{log}{if \code{TRUE}, returns probabilities on the log scale.}

\item{min.phi}{the minimum bite size in terms of proportion leaf herbivory. Defaults to 0.005 (0.5\%).}

\item{max.phi}{the maximum bite size in terms of proportion leaf herbivory Defaults to 1 (100\%).}

\item{a}{the combined allometric scaling coefficient. Defaults to 14/9.}

\item{k.fft.limit}{the maximum number of convolutions performed by FFT. For k convolutions above this limit, a Gaussian approximation is used.}

\item{parallel}{if \code{TRUE} (default is \code{FALSE}), convolve \code{fft.vec} in parallel. Automatically turned off if \code{length(ff.vec)} is less than 50000, as no computational efficiency would be gained.}

\item{phi.T.index}{a vector of integers indicating the location where phi.T most closely match the grid of probabilities. If \code{NULL} (default), the indices are calculated using \code{phi.T}, \code{min.phi}, \code{max.phi}, and \code{a}. Supplying a pre-calculated index vector saves a lot of computation when the function is used repeatedly.}
}
\value{
a \code{length(phi)} X \code{length(k)} matrix of conditional probabilities.
}
\description{
Internal function of \code{dalloT()} used to estimate the conditional probability matrix \eqn{P(\phi_T|k)} via FFT or Gaussian approximation.
}
