% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{ad.test}
\alias{ad.test}
\alias{hessian}
\alias{KL}
\title{Other handy exported functions}
\usage{
ad.test(...)

hessian(...)

KL(x, test.na = TRUE, unit = "log", est.prob = NULL)
}
\arguments{
\item{...}{additional arguments passed to the exported function}

\item{test.na, unit, est.prob, x}{See \code{philentropy::KL()}.}
}
\value{
see documentation of exported functions.
}
\description{
Essentially a wrapper for certain useful functions exported form other packages.
}
