% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{Gini}
\alias{Gini}
\alias{Skew}
\alias{Kurt}
\title{Summary statistics of x}
\usage{
Gini(x, na.rm = FALSE, fast = TRUE, unbiased = TRUE, ...)

Skew(x, weights = NULL, na.rm = FALSE, method = 1, fast = TRUE, ...)

Kurt(x, weights = NULL, na.rm = FALSE, method = 1, fast = TRUE, ...)
}
\arguments{
\item{x}{a numeric vector for which the summary statistic is calculated}

\item{na.rm}{if \code{TRUE} (default is \code{FALSE}), the missing values are removed}

\item{fast}{if \code{TRUE}, \code{Rfast} is used to do the calculation. It's much faster but handles fewer edge cases.}

\item{unbiased}{If \code{TRUE} (default), the unbiased Gini coefficient is calculated.}

\item{...}{additional arguments passed to \code{DescTools}}

\item{method, weights}{\code{method} and \code{weights} argument passed to \code{DescTools}.}
}
\value{
a numeric value
}
\description{
Summary statistics of x calculated using \code{Rfast} or \code{DescTools}. For \code{Kurt()}, \code{DescTools} returns the excess Kurtosis, but \code{herbivar::Kurt()} returns the raw kurtosis.
}
