% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{fg_threshold}
\alias{fg_threshold}
\title{Threshold Foreground With Automatic Threshold Selection}
\usage{
fg_threshold(
  img,
  fg.thresh = "auto",
  fg.adjust = 1,
  approx.res = 10000,
  km = c(1L, 2L),
  fg.thr = NULL
)
}
\arguments{
\item{img}{a cimg object}

\item{fg.thresh}{a threshold for turning the cimg pixels into a selected pixset. Can be either numeric, a string, or "auto". When set to "auto", the threshold is automatically selected by the internal function \code{threshold2()}.}

\item{fg.adjust}{used to adjust the automatic threshold. If the auto-threshold is at k, the effective threshold will be at \code{fg.adjust} * k.}

\item{approx.res}{the number of pixels used in threshold approximation. Default is 10000.}

\item{km}{an integer vector of length two. The first value selects the the boundary rank in increasing means identified by kmeans clustering (defaults to 1, selecting the boundary that separates the two clusters with the lowest mean). The second value selects the k number of clusters to form (defaults to 2). Ignored when not auto-thresholding.}

\item{fg.thr}{an internal argument to pass calculated threshold to reduce redundant computation. Must be a numeric value.}
}
\value{
a pixset with the selected pixels
}
\description{
Internal function of \code{clean_leaf()}. Basically acts as a wrapper for \code{threshold()} with added threshold reporting and error handling. When foreground threshold is selected, either manually or automatically, above 100\%, the threshold is reduced by 10\% recursively up to five times until foreground threshold is less than 100\%. Automatically selected threshold by \code{threshold()}that is below 40\% maybe too low and is increased by 30\% automatically.
}
