% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_ML_support.R
\name{fit_bite_size}
\alias{fit_bite_size}
\title{Fit A 'Bite Size' Distribution To Data}
\usage{
fit_bite_size(
  object,
  family = c("allo", "allo_a", "allo_M", "tlnorm", "beta", "kumar", "tpareto", "cb"),
  min.phi = NA,
  method = c("BFGS", "nlminb", "Nelder-Mead"),
  IC = "AICc"
)
}
\arguments{
\item{object}{A vector of proportion 'bite sizes' between zero and one, not inclusive. If an object of class 'split_herb' is supplied, the \code{hole_prop} vector will be extracted.}

\item{family}{A vector of distributions to fit to a vector of 'bite sizes'. Valid options are "allo", "allo_a", "allo_M", "tlnorm", "beta", "kumar", "tpareto", and "cb". See details. If "all", all available families will be selected.}

\item{min.phi}{The minimum proportion 'bite size'. If \code{NA} (default), the minimum non-zero value in the data is used. If \code{NA} and \code{object} supplied is of class 'split_herb', the \code{min_prop} is extracted from the \code{object}.}

\item{method}{Optimizer used to find the maximum likelihood estimates. Valid options are "Nelder-Mead", "BFGS" (default), and "nlminb".}

\item{IC}{An information criteria used to rank the fitted distributions. Default is "AICc".}
}
\value{
A object of class 'bite_size_fit' and 'list'

\code{models}:
A named list of model fit details, including MLE values, SE, log likelihood, and convergence code.

\code{settings}:
A named list of function inputs, including the original vector of data, \code{min.phi}, \code{method}, \code{family}, and \code{herbivar.version}.

\code{summary}:
A matrix array of ordered IC values of fitted models.
}
\description{
Fit multiple distributions to a vector of data using maximum likelihood.
}
\details{
\subsection{allo}{

Neutral 'bite size' distribution with zero degree of freedom.
\deqn{P(\phi| \phi_M = 1, \phi_m = min.phi, \alpha = \frac{14}{9}) = \frac{1-\alpha}{\phi_M^{1-\alpha} - \phi_m^{1-\alpha}} \phi^{-\alpha}.}
}

\subsection{allo_a}{

Neutral 'bite size' distribution with one degree of freedom, allowing \eqn{\alpha} to be fitted.
\deqn{P(\phi| \phi_M = 1, \phi_m = min.phi, \alpha) = \frac{1-\alpha}{\phi_M^{1-\alpha} - \phi_m^{1-\alpha}} \phi^{-\alpha}.}
}

\subsection{allo_M}{

Neutral 'bite size' distribution with one degree of freedom, allowing \eqn{\phi_M} to be fitted. The MLE of \eqn{\phi_M} is always the maximum observed value in the data (Aban et al. 2006). A convergence code of 2 is returned.
\deqn{P(\phi| \phi_M, \phi_m = min.phi, \alpha = \frac{14}{9}) = \frac{1-\alpha}{\phi_M^{1-\alpha} - \phi_m^{1-\alpha}} \phi^{-\alpha}.}
}

\subsection{tlnorm}{

Truncated log-normal distribution with two degrees of freedom.
\deqn{P(\phi) = \frac{g(\phi, \mu, \sigma)}{G(\phi = 1; \mu ,\sigma)} }
The log-normal PDF is defined as
\deqn{ g(\phi; \mu , \sigma) = \frac{1}{\phi \sigma \sqrt{2 \pi}} \exp{(-\frac{(\ln{\phi}-\mu)^2}{2\sigma^2})} }
and the log-normal CDF is \eqn{G(\phi = 1; \mu ,\sigma)}, \eqn{\mu} is the mean, \eqn{\sigma} is the standard deviation.
}

\subsection{beta}{

Beta distribution with two degrees of freedom.
\deqn{P(\phi) = \frac{\Gamma(a+b)}{\Gamma(a)\Gamma(b)}\phi^{a-1}(1-\phi)^{b-1}}
\eqn{a} and \eqn{b} are the two shape parameters.
}

\subsection{kumar}{

Kumaraswamy distribution with two degrees of freedom.
\deqn{P(\phi) = ab\phi^{a-1}(1-\phi^{a-1})(1-\phi^{a})^{b-1}}
\eqn{a} and \eqn{b} are the two shape parameters.
}

\subsection{tpareto}{

Truncated Pareto distribution with two degrees of freedom. This is equivalent to the neutral 'bite size' distribution with the parameters \eqn{\phi_m} and \eqn{a} fitted to data.
\deqn{P(x) = \frac{a b^a}{\phi^{a+1}}\frac{1}{1 - (\frac{b}{1})^a},}.
\eqn{a} is the exponent of the Pareto distribution and \eqn{b} is the minimum value of \eqn{\phi}.
}

\subsection{cb}{

Continuous Bernoulli distribution with one degree of freedom.
\deqn{P(\phi) = C(\lambda) \lambda^\phi (1 - \lambda)^{1-\phi},}
\eqn{C(\lambda)} is defined as
\deqn{C(\lambda) = 2}
if \eqn{\lambda = \frac{1}{2}}, otherwise,
\deqn{C(\lambda) = \frac{2 \tanh^{-1}(1-2\lambda)}{1-2\lambda}.}

\eqn{\lambda} is the rate parameter.
}
}
\references{
Aban, I. B., M. M. Meerschaert, and A. K. Panorska. 2006. Parameter Estimation for the Truncated Pareto Distribution. Journal of the American Statistical Association 101:270–277.
Loaiza-Ganem, G., and J. P. Cunningham. 2019. The continuous Bernoulli: fixing a pervasive error in variational autoencoders. Page Advances in Neural Information Processing Systems. Curran Associates, Inc.
}
