% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_ML_support.R
\name{fit_generic_null}
\alias{fit_generic_null}
\title{Fit Non-Neutral Generic Null Models Using Maximum Likelihood Estimation}
\usage{
fit_generic_null(
  data.vec,
  family = c("htlnorm", "zoibeta"),
  init = 0,
  lower = -Inf,
  upper = Inf,
  method = c("Nelder-Mead", "BFGS", "L-BFGS-B", "nlminb", "nlm"),
  id = NULL,
  ...
)
}
\arguments{
\item{data.vec}{A vector of numeric data. \code{NA}s are ignored. If the whole vector does not contain any non-zero values or contains values outside of \eqn{[0,1]}, the function would throw an error.}

\item{family}{The generic null distribution. Valid options are "htlnorm" for hurdle truncated log-normal distribution, and "zoibeta" for zero-one-inflated beta distribution. see \code{?dzoibeta()} and \code{?dhtlnorm()}}

\item{init}{Initial value for optimizer to start. Defaults to 0.}

\item{lower, upper}{A numeric vector indicating the bounds of each estimated variable for the "L-BFGS-B" and "Brent" method. Defaults to \code{-Inf} and \code{Inf}.}

\item{method}{The method of maximum likelihood estimation algorithm. Acceptable methods are "Nelder-Mead" (Default), "BFGS", "L-BFGS-B", "Brent", "nlminb", and "nlm". Defaults to "Brent" for one dimensional optimization. For details see \code{?stats::optim()} and \code{?stats::nlminb()}.}

\item{id}{A character string supplied for book keeping purposes. Default is \code{NULL}. Useful for when storing a large number of "generic_null_fit" objects in a list.}

\item{...}{Additional arguments that are passed to the optimizer functions \code{optim()} or \code{nlminb()}.}
}
\value{
An 'generic_null_fit' object with the following slots:

\code{theta.names}: A vector of character string naming the fitted parameters

\code{par}: the value of fitted parameters on the transformed scale in the order of theta.names

\code{se}: standard error of fitted parameters on the transformed scale. Quadratic approximation of the standard error will be implemented in the future.

\code{loglik}: log likelihood of the data given the parameter combinations

\code{hessian}: The Hessian matrix

\code{iters}: The number of times the likelihood function or its gradient is evaluated.

\code{convergence}: 0 indicates successful convergence. Error codes are passed from the optimizer function. 1 indicate that the iteration limit has been reached and convergence has failed. 2 indicate that the Hessian matrix is non-positive definite.

\code{method}: Method of optimization

\code{init}: A vector of the used initial values for optimization

\code{data}: The data to which the model is fitted to

\code{param.val.trans}: A vector of functions that transform each estimated parameter value in the optimization process (the optimizer find the values on the transformed scale).

\code{id}: name of model supplied via the id argument

\code{n}: the sample size

\code{df}: the degrees of freedom of the model
}
\description{
Estimate unknown parameters in non-neutral generic null herbivory models from a vector of observed herbivory data using Maximum Likelihood Estimation (MLE).
}
