% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FOSM}
\alias{FOSM}
\title{First Order Second Moment Method}
\usage{
FOSM(x, var, trans)
}
\arguments{
\item{x}{The expectation of the random variable}

\item{var}{The variance of the random variable}

\item{trans}{A function by which the random variable is transformed}
}
\value{
a numeric value
}
\description{
Compute the variance of a transformed random variable using first order Taylor's approximation (Delta method).
}
\details{
The first order second moment method can be used to approximate the variance of a transformed random variable provided that we know the first derivative of the transformation function, the mean, and variance of the random variable:
\deqn{\mathbb{Var}[f(X)] \approx f'(\mathbb{E}[X])^2 \mathbb{Var}[X]}
}
