% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{gen_mean}
\alias{gen_mean}
\title{Generalized Mean}
\usage{
gen_mean(x, p, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of numeric data}

\item{p}{a numeric value of length 1 choosing the type of mean computed.}
}
\value{
a numeric value of length 1
}
\description{
Calculate generalized mean for a vector \code{x}. If \eqn{p = 0}, the function returns the geometric mean. If \eqn{p = -1}, the function returns the harmonic mean. If \eqn{p = 1} or \eqn{p = 2}, the arithmetic mean or the root mean square (quadratic mean) is returned. If \eqn{p = \infty} or \eqn{p = -\infty}, the function returns the maximum or minimum respectively. For other values of \eqn{p}, the power mean is returned. For large values of \eqn{|p|}, the function can be numerically unstable.
}
