% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{get_data_sim}
\alias{get_data_sim}
\title{Generate Neutral Predictions From Fitted Model}
\usage{
get_data_sim(
  allo.fit.list,
  n.sim = NULL,
  digits = 2,
  return.obs = TRUE,
  obs.raw = FALSE,
  new.param = c(mean.phi.T = NA, min.phi = NA, max.phi = NA, a = NA)
)
}
\arguments{
\item{allo.fit.list}{A list of "allo_herb_fit" objects.}

\item{n.sim}{A numeric value indicating number of points to draw from the neutral distribution for each "allo_herb_fit" object. If \code{NULL}, the original sample size of the data will be used.}

\item{digits}{A numeric value indicating the number of digits to round data values to}

\item{return.obs}{A logical value indicating whether to return observed data as a list in the out put along with the predicted draws}

\item{obs.raw}{A logical value indicating whether to round the observed data; ignored if \code{return.obs} is set to \code{FALSE}.}

\item{new.param}{A vector of named values used to make new predictions. If a parameter is set to \code{NA}, the parameter value is extracted from the "allo_herb_fit" object.}
}
\value{
A list of list of vectors of numeric values.
}
\description{
This is a convenience function that generates predicted distribution from a list of fitted models
}
