% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{get_dist_test_sim}
\alias{get_dist_test_sim}
\title{Compare whether two samples are drawn from the same distribution En Mass}
\usage{
get_dist_test_sim(
  fit.list,
  test = c("ks", "kl", "ad", "chisq"),
  nboot = 1,
  n.sim = NULL,
  digits = 2,
  silent = FALSE,
  ...
)
}
\arguments{
\item{fit.list}{A list of fitted objects that is passed to \code{get_data_sim()}}

\item{test}{A vector of character string indicating which methods to use to compare the two samples. "ks" performs the KS test using \code{stats::ks.test()}. "ad" performs the AD test using \code{kSamples::ad.test()}. "kl" computes the KL divergence using \code{philentropy::KL()}. The unit of the KL divergence is "log" by default. "chisq" performs the Chi-square test using \code{stats::chisq.test()}.}

\item{nboot}{Number of times to repeat the simulation. That is, drawing samples from each fitted object then comparing the draws to observed data.}

\item{n.sim}{A numeric value indicating number of points to draw from the neutral distribution for each "allo_herb_fit" object. If \code{NULL}, the original sample size of the data will be used.}

\item{digits}{An integer indicating the number of digits the samples should be rounded to before doing the calculations. This can be important when rounding gets rid of some sampling artifacts in the data.}

\item{silent}{If \code{TRUE}, the function will not print a progress counter.}

\item{...}{additional parameters passed to \code{get_data_sim()}}
}
\value{
A length(fit.list) X length(test statistics) X nboot array.
}
\description{
A batch processing wrapper for \code{compare_dist()} and \code{get_data_sim()}
}
