% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{get_images}
\alias{get_images}
\title{Download images from URL}
\usage{
get_images(
  x,
  save_dir,
  enforce_file_extension = "jpg",
  confirm = TRUE,
  skip = NULL
)
}
\arguments{
\item{x}{a named vector with URLs as the values and file name (without the extension) as the name of the values.}

\item{save_dir}{a path to where the files will be saved}

\item{enforce_file_extension}{File extension to be enforced (default is \code{"jpg"}). Encountering unexpected file types will result in an error.}

\item{confirm}{if \code{TRUE} (default), a prompt will appear to check if the user wants to continue with the download.}

\item{skip}{a vector of character string of file names in \code{x} that should be skipped.}
}
\description{
Download images from a named vector of URL character string.
}
