% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model_ML_support.R
\name{get_lambda}
\alias{get_lambda}
\title{Calculate Attack Rate From Unlimited Mean Cumulative Proportion Herbivory}
\usage{
get_lambda(mean.phi.T, min.phi = 0.005, max.phi = 1, a = 14/9)
}
\arguments{
\item{mean.phi.T}{The mean herbivory of the distribution when herbivores are not plant limited.}

\item{min.phi}{the minimum bite size in terms of proportion leaf herbivory. Defaults to 0.005 (0.5\%).}

\item{max.phi}{the maximum bite size in terms of proportion leaf herbivory Defaults to 1 (100\%).}

\item{a}{the combined allometry scaling coefficient. Defaults to 14/9.}
}
\value{
a vector of numeric values
}
\description{
Convenience function that reparameterizes unlimited mean cumulative proportion herbivory \eqn{\overline{\phi_T'}} as attack rate \eqn{\lambda}.
}
\details{
\deqn{\lambda = \frac{\overline{\phi_{T}'}}{\overline{\phi}},} where
\deqn{\overline{\phi} = \frac{1-\alpha}{2-\alpha} \frac{\phi_M^{2-\alpha} - \phi_m^{2-\alpha}}{\phi_M^{1-\alpha} - \phi_m^{1-\alpha}}.}
}
