% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model_ML_support.R
\name{get_phi_bar}
\alias{get_phi_bar}
\title{Calculate Mean Proportion Bite Size}
\usage{
get_phi_bar(min.phi = 0.005, max.phi = 1, a = 14/9)
}
\arguments{
\item{min.phi}{the minimum bite size in terms of proportion leaf herbivory. Defaults to 0.005 (0.5\%).}

\item{max.phi}{the maximum bite size in terms of proportion leaf herbivory Defaults to 1 (100\%).}

\item{a}{the combined allometric scaling coefficient. Defaults to 14/9.}
}
\value{
a numeric value of the mean
}
\description{
Calculate the theoretical mean of the proportion bite size distribution
}
\details{
The mean of the bite size distribution can be found via:

When \eqn{\phi \neq 1,2}
\deqn{\overline{\phi} = \frac{1-\alpha}{2-\alpha} \frac{\phi_M^{2-\alpha} - \phi_m^{2-\alpha}}{\phi_M^{1-\alpha} - \phi_m^{1-\alpha}}}

when \eqn{\phi = 1}
\deqn{\overline{\phi} = \frac{1}{2-\alpha} \frac{\phi_M^{2-\alpha} - \phi_m^{2-\alpha}}{\ln{\phi_M} - \ln{\phi_m}}}

when \eqn{\phi = 2}
\deqn{\overline{\phi} = (1-\alpha) \frac{\ln{\phi_M} - \ln{\phi_m}}{\phi_M^{1-\alpha} - \phi_m^{1-\alpha}}}
}
