% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model_ML_support.R
\name{get_phi_var}
\alias{get_phi_var}
\title{Calculate Variation of Proportion Bite Size}
\usage{
get_phi_var(min.phi = 0.005, max.phi = 1, a = 14/9)
}
\arguments{
\item{min.phi}{the minimum bite size in terms of proportion leaf herbivory. Defaults to 0.005 (0.5\%).}

\item{max.phi}{the maximum bite size in terms of proportion leaf herbivory Defaults to 1 (100\%).}

\item{a}{the combined allometric scaling coefficient. Defaults to 14/9.}
}
\value{
a numeric value of the variance
}
\description{
Calculate the theoretical variance of the proportion bite size distribution
}
\details{
The variance of the bite size distribution can be found via:

When \eqn{\phi \neq 1,3}
\deqn{\mathbb{V}ar[\phi] = \frac{1-\alpha}{3-\alpha} \frac{\phi_M^{3-\alpha} - \phi_m^{3-\alpha}}{\phi_M^{1-\alpha} - \phi_m^{1-\alpha}} - \overline{\phi}^2}

when \eqn{\phi = 1}
\deqn{\mathbb{V}ar[\phi] = \frac{1}{3-\alpha} \frac{\phi_M^{3-\alpha} - \phi_m^{3-\alpha}}{\ln{\phi_M} - \ln{\phi_m}} - \overline{\phi}^2}

when \eqn{\phi = 3}
\deqn{\mathbb{V}ar[\phi] = (1-\alpha) \frac{\ln{\phi_M} - \ln{\phi_m}}{\phi_M^{1-\alpha} - \phi_m^{1-\alpha}} - \overline{\phi}^2}
}
