% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{grow_edge}
\alias{grow_edge}
\title{Add One Row of Pixel to Edge}
\usage{
grow_edge(object, background = NA)
}
\arguments{
\item{object}{a 'cimg' object}

\item{background}{the value appended to the edge of the image. Default is \code{NA}. Only a single value is supported at the moment. So for colored images, all color channels will be set the same value.}
}
\value{
A 'cimg' object
}
\description{
Add one row of pixel with value 'background' (default is \code{NA}) to the edge of the image. Useful for image processing when the leaf is touching the edge of the image, which prevents the leaf boarder from being recognized properly.
}
