% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hellinger_trans}
\alias{hellinger_trans}
\title{Hellinger Transformation}
\usage{
hellinger_trans(x)
}
\arguments{
\item{x}{a community matrix where each column is a species and each row is a site}
}
\value{
a transformed community matrix
}
\description{
The Hellinger transformation for a community matrix where the relative species abundance at each site (row) is square root transformed. If no species is observed at a site, the function returns a zero for the site.
}
