% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{herb_data_example}
\alias{herb_data_example}
\title{Herbivory data example}
\format{
\subsection{\code{herb_data_example}}{

A list of two vectors
\describe{
\item{field}{Field survey data of \emph{H. matronalis} leaf proportion chewing damage}
\item{lab}{Proportion chewing damage on agar blocks in a lab setting (limited regional herbivore pool)}
}
}
}
\usage{
herb_data_example
}
\description{
Real proportion herbivory data from a field survey of \emph{Hesperis matronalis} and agar artificial diet blocks fed to caterpillar a single \emph{Spodoptera exigua} caterpillar. The minimum non-zero proportion herbivory is 0.5\%.
}
\keyword{datasets}
