% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbivar-package.R
\docType{package}
\name{herbivar-package}
\alias{herbivar}
\alias{herbivar-package}
\title{herbivar: Quantitative Tools for Variance Explicit Ecology and Herbivory}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package makes available several methods used for analyzing variability in ecological data. Currently, it supports fitting a neutral herbivory model to data using Maximum Likelihood Estimation (MLE). It also contains several helper functions for comparing distributions and assessing model fit. In addition, using the 'imager' infrastructure, the package also supports automatic processing of leaf images, mostly in the context of leaf scans, that can retrieve rich within-leaf herbivory data. The package is integrated with the 'spatstat' packages that allows for spatial analysis of these data.
}
\author{
\strong{Maintainer}: Vincent Pan \email{vsbpan@gmail.com} (\href{https://orcid.org/0000-0001-9892-7805}{ORCID})

}
\keyword{internal}
