% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_management.R
\name{herbivar.version}
\alias{herbivar.version}
\title{Prints herbivar version}
\usage{
herbivar.version(silent = FALSE)
}
\arguments{
\item{silent}{if \code{TRUE}, no message will be displayed, but returned invisibly}
}
\value{
a character string
}
\description{
Prints herbivar version
}
