% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{hole_perimeter}
\alias{hole_perimeter}
\title{Calculate Perimeter}
\usage{
hole_perimeter(mat, non_leaf_border = FALSE, px.size = NA)
}
\arguments{
\item{mat}{a cimg, pixset, matrix, or array}

\item{non_leaf_border}{if \code{TRUE} (default is \code{FALSE}), return perimeter that includes non-leaf borders (edges that separate eaten leaf area and non-leaf area).}

\item{px.size}{value passed to \code{px_size_calc()}. When set to \code{NA} (default), the 'px.size' attribute is extracted from the supplied \code{mat} object.}
}
\value{
a vector of the perimeter in the number of pixels and millimeters. The object must have a "px.size" attribute or the \code{px.size} argument must be defined to return values for mm perimeter
}
\description{
Find the perimeter of selected pixels (must have value 1) including or excluding non-leaf borders.
}
