% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{Image2array}
\alias{Image2array}
\title{Convert Array to Class 'Image'}
\usage{
Image2array(x, imager_class = TRUE)
}
\arguments{
\item{x}{An 'Image' object}

\item{imager_class}{If \code{TRUE} (default), convert the array into a cimg or pixset depending on whether the array elements are numeric or logical.}
}
\value{
A four dimensional array (as in the format in the package \code{imager}), a cimg, or pixset object.
}
\description{
Convert an object of class 'Image' from the package \code{EBImage} to an array formatted as cimg or pixset.
}
