% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{image_example}
\alias{image_example}
\title{Load Example Leaf Image}
\usage{
image_example(type = c("raw", "processed"))
}
\arguments{
\item{type}{a character string indicating the example image to be loaded}
}
\value{
an object of class 'cimg'
}
\description{
A convenience function for loading example image built in the package.
}
