% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{immask}
\alias{immask}
\title{Mask Unselected Pixels}
\usage{
immask(object, pixset, background = NA)
}
\arguments{
\item{object}{a cimg or array for which the values of unselected pixel are set}

\item{pixset}{a pixset object defining the set of pixels to be retained}

\item{background}{a numeric vector of equal length as the number of color spectra, a character string to be parsed by \code{col2rgb()}, or \code{NA} (default).}
}
\value{
a masked object
}
\description{
Set pixels in an image not selected in the pixset as an user defined value
}
\examples{
img <- image_example()
px <- threshold2(img,thr = 0.3)
plot(px)

immask(img,px,"hotpink") \%>\% plot()
immask(img, px) \%>\% plot()

immask(grayscale(img),B(px), 0) \%>\% plot()

}
