% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv.hsin}
\alias{inv.hsin}
\title{Inverse Hyperbolic Sine Transformation}
\usage{
inv.hsin(x)
}
\arguments{
\item{x}{a numeric vector to be transformed}
}
\value{
a transformed numeric vector
}
\description{
The inverse hyperbolic sine transformation that can be used in place of the log transformation when the vector contains zero.
}
\references{
Burbidge JB, Magee L, Robb AL (1988) Alternative Transformations to Handle Extreme Values of the Dependent Variable. Journal of the American Statistical Association 83:123–127.
}
