% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{is.margin}
\alias{is.margin}
\title{Check Margin}
\usage{
is.margin(mat)
}
\arguments{
\item{mat}{a cimg, pixset, matrix, or array}
}
\value{
an atomic logical value
}
\description{
Check if the selected pixels (must have value 1) is on the leaf margin (i.e. bordering \code{NA}s).
}
