% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{J.index}
\alias{J.index}
\title{Calculate Aggregation Coefficient}
\usage{
J.index(x = NULL, mean = NULL, var = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of numeric data}

\item{mean}{mean of data. Ignored if \code{x} is supplied}

\item{var}{variance of data. Ignored if \code{x} is supplied}

\item{na.rm}{a logical value indicating whether to drop NA values}
}
\value{
A numeric value
}
\description{
calculate the aggregation \eqn{J} for a given mean and variance or vector of data
}
\details{
The aggregation coefficient \eqn{J} is calculated as
\deqn{J = \frac{\sigma^2 - \mu}{\mu^2}}
\eqn{J = 0} corresponds to a possion distribution (random)
\eqn{J < 0} corresponds to a more even distribution (underdispersed)
\eqn{J > 0} corresponds to a more aggregated distribution (overdispersed)
@export
}
