% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-linear averaging simulation.R
\name{JE}
\alias{JE}
\alias{JE2}
\title{Non-linear Averaging Simulation}
\usage{
JE(model, x, RV.name = NULL, nboot = 1L, summarise = TRUE, plot = FALSE)

JE2(
  model,
  x,
  RV.name = NULL,
  nboot = 1L,
  summarise = TRUE,
  plot = FALSE,
  spaghetti = TRUE
)
}
\arguments{
\item{model}{A function or a fitted model that can be handled by the package 'insight'.}

\item{x}{A vector of numeric value used for the simulation.}

\item{RV.name}{An optional character value naming the predictor if \code{model} is a fitted model, telling the model which predictor to use for generating predictions.}

\item{nboot}{An integer value indicating the number of bootstraps for estimating the uncertainty in the Jensen's effect due to the uncertainty of the function. Only relevant for when \code{model} is fitted. Provided \code{model} is treated as exact when it is a function.}

\item{summarise}{If \code{TRUE}, the output will be summarized. Otherwise, the raw simulated draws will be returned.}

\item{plot}{If \code{TRUE}, a plot will be generated.}

\item{spaghetti}{If \code{TRUE} (default), a spaghetti plot will be generated.}
}
\value{
\code{JE} generates a vector of simulated Jensen's effect.
\code{JE2} generates a data.frame of simulated Jensen's effect potential (second derivative or acceleration) of each unique value of \code{x}.
}
\description{
Estimate Jensen's effect via direct simulation or Jensen's effect potential via second order Taylor approximation.
}
\details{
The Jensen's effect \eqn{J} at \eqn{\mathbb{E}[X])} is calculated exactly as:

\deqn{J = \mathbb{E}[f(X)] - f(\mathbb{E}[X]).}

Jensen's effect is related to the local acceleration as we can see from the second order Taylor approximation:

\deqn{J(x) = \frac{1}{2}\frac{d^2}{dx^2} f(x) \mathbb{Var}[X] + \mathcal{O}(x^3).}

Therefore, Jensen's effect potential can be expressed in terms of the local second derivative of \eqn{f(x)}, where multiplication by \eqn{\frac{1}{2}\mathbb{Var}[X]} gives us an approximation of \eqn{J(x)}.
}
\examples{
f <- function(x){
 x^3 - 0.5*x^2 + x
}

x <- rnorm(100, 0, 1)
y <- f(x)
plot(y~x)

m <- lm(y ~ x + I(x^3))
y.mpred <- predict(m)
plot(y.mpred ~ x)

JE(f, x, plot = TRUE)
JE(m, x, "x", nboot = 10, plot = TRUE)

JE2(f,x, "x", nboot = 10, plot = TRUE)
JE2(m,x, "x", nboot = 10, plot = TRUE)

}
