% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{lac}
\alias{lac}
\title{Lorenz Asymmetry Coefficient}
\usage{
lac(x, n = NULL, interval = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of non-negative numeric values}

\item{n}{a vector of frequencies the same length as \code{x}. If \code{NULL} (default), the frequencies are assumed to be 1 (as in a vector of untabulated raw data).}

\item{interval}{a logical value indicating whether to return an interval when there are x values exactly equal to the mean. If \code{FALSE} (default), the midpoint is returned.}

\item{na.rm}{a logical value indicating whether to drop \code{NA} values}
}
\value{
a single numeric value if \eqn{x_i \neq \hat\mu} for all values of \code{x} or if \code{interval = FALSE}. Otherwise, a vector of length two is returned.
}
\description{
Calculate Lorenz Asymmetry Coefficient (also known as \eqn{S}) for a vector of data. The statistic identifies which quantile of \code{x} contribute most to the total inequality.
}
\details{
The Lorenz Asymmetry Coefficient \eqn{S} is a complementary statistic to the Gini coefficient that describes the Lorenz curve. Whereas the Gini coefficient is a measure of inequality, the asymmetry coefficient is is a measure of which half of the \code{x} quantile contribute most to the inequality. If \eqn{S = 1}, the Lorenz curve is symmetrical. If \eqn{S < 1}, the point where the Lorenz curve is parallel to the line of equality is below the axis of symmetry. This means that the lower half of the population contribute most to the inequality. If \eqn{S > 1}, the point where the Lorenz curve is parallelt to the line of equality is above the axis of symmetry. This means that the upper half of the population contribute most to the inequality. See example plot.

For a vector of ordered non-negative value \eqn{(x_1, x_2, ..., x_m, x_{m+1}, ..., x_n)}, the sample Lorenz Asymmetry is defined as:
\deqn{S = F(\hat\mu) + L(\hat\mu)}

where
\deqn{\delta = \frac{\hat\mu - x_m}{x_{m+1} - x_m},}

\deqn{F(\hat\mu) = \frac{m + \delta}{n},}

\deqn{L(\hat\mu) = \frac{L_m + \delta x_{m+1}}{L_m}.}

\eqn{\hat\mu} is the mean of \eqn{x}, \eqn{n} is the sample size, \eqn{m} is the number of elements where \eqn{x_m < \hat\mu}, and \eqn{L_i = \sum_{j=1}^i x_j}.

If one or more values of \eqn{x_i = \hat\mu}, the closed interval is calculated instead:

\deqn{[ \frac{m}{n} + \frac{L_m}{L_n}, \frac{m + a}{n} + \frac{L_{m+a}}{L_n} ],}
where \eqn{a} is the number of elements of the vector \eqn{x} that satisfies \eqn{x_i = \hat\mu}.
}
\note{
Part of the code for this function is adapted from Achim Zeileis's \code{ineq::Lasym()} version 0.2-13.
}
\examples{
x <- rbeta(1000,1,0.2)
hist(x)
plot(lorenz_curve(x))
abline(a=1,b=-1) # Line of symmetry
lac(x) # S = 0.48

x2 <- rallo(1000,a=3)
hist(x2)
plot(lorenz_curve(x2))
abline(a=1,b=-1) # Line of symmetry
lac(x2) # S = 1.24

}
\references{
Damgaard, C., and J. Weiner. 2000. Describing Inequality in Plant Size or Fecundity. Ecology 81:1139–1142.
}
