% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{leaf_area}
\alias{leaf_area}
\title{Calculate Leaf Area}
\usage{
leaf_area(object, px.size = NA)
}
\arguments{
\item{object}{an object of type 'split_herb', 'matrix', 'array ', 'cimg', or 'pixset'.}

\item{px.size}{value passed to \code{px_size_calc()}. When set to \code{NA} (default), the 'px.size' attribute is extracted from the supplied \code{mat} object.}
}
\value{
a vector of the number of pixels and milometer squared area of leaf area. The object must have a "px.size" attribute or the \code{px.size} argument must be defined to return values for mm2 leaf area.
}
\description{
calculate leaf area pixels and mm2. Non-leaf parts of the image matrix must be indicated by \code{NA}s.
}
