% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{leaf_herb}
\alias{leaf_herb}
\title{Calculate Leaf Herbivory}
\usage{
leaf_herb(
  object,
  type = c("proportion", "percent", "pixels", "px", "mm2", "real"),
  px.size = NA
)
}
\arguments{
\item{object}{an object of type 'split_herb', 'matrix', 'array ', 'cimg', or 'pixset'.}

\item{type}{a vector of character string indicating what values to return. Partial string matching supported. "proportion" and "percent" returns relative proportion herbivory. "pixels" and "px" returns the number of pixels eaten. "mm2" and "real" returns the leaf area eaten in mm2.}

\item{px.size}{value passed to \code{px_size_calc()}. When set to \code{NA} (default), the 'px.size' attribute is extracted from the supplied \code{object}.}
}
\value{
a vector of proportion, number of pixels, and milometer squared area of herbivory. The object must have a "px.size" attribute or the \code{px.size} argument must be defined to return values for mm2 leaf area removed.
}
\description{
calculate leaf herbivory in proportion, pixels, and mm2. Non-leaf parts, remaining leaf parts, and eaten leaf parts of the image matrix must be indicated by \code{NA}s, 0, and 1 (or 3) respectively.
}
