% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{leaf_length}
\alias{leaf_length}
\title{Find Major Axis Length of Leaf}
\usage{
leaf_length(object, px.size = NA)
}
\arguments{
\item{object}{an object of type 'split_herb', 'matrix', 'array ', 'cimg', or 'pixset'.}

\item{px.size}{value passed to px_size_calc(). When set to NA (default), the 'px.size' attribute is extracted from the supplied object}
}
\value{
a vector of the number of pixels and milometer maximum distance between leaf border pixels. he object must have a "px.size" attribute or the px.size argument must be defined to return values for mm2 distance.
}
\description{
Takes a cropped image and find the maximum euclidean distance between leaf borders. Some caveats apply. For instance, if the leaf is not straight, then the distance would be off.
}
