% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model_ML_support.R,
%   R/null_model_ML_support.R
\name{logLik.allo_herb_fit}
\alias{logLik.allo_herb_fit}
\alias{logLik.generic_null_fit}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{allo_herb_fit}(object, ...)

\method{logLik}{generic_null_fit}(object, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{...}{additional arguments}
}
\value{
a 'logLik' object
}
\description{
extract log-likelihood
}
