% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{lorenz_curve}
\alias{lorenz_curve}
\title{Calculate Lorenz Curve}
\usage{
lorenz_curve(x, n = NULL, na.rm = FALSE, nboot = 100, return.array = FALSE)
}
\arguments{
\item{x}{a vector of non-negative numeric data}

\item{n}{a vector of frequencies the same length as \code{x}. If \code{NULL} (default), the frequencies are assumed to be 1 (as in a vector of untabulated raw data).}

\item{na.rm}{a logical value indicating whether to drop \code{NA} values}

\item{nboot}{an integer value indicating the number of bootstraps to perform to estimate the uncertainty (default to 100). If set below 0 or \code{FALSE}, no bootstraps will be performed. If set to \code{TRUE}, 100 bootstraps will be performed.}

\item{return.array}{a logical value indicating whether to save the bootstrapped Lorenz Curves. Default is \code{FALSE}. Must set to \code{TRUE} for spaghetti plot in \code{plot.lc()}.}
}
\value{
an object of class "lc" and "list" with three slots:

\code{lc}: a data.frame of the exact Lorenz Curve calculated from the supplied data

\code{lc_boot_summary}: a data.frame of the mean and standard error of the Lorenz Curve. If no bootstrap is performed, \code{NULL} is returned.

\code{boot_array}: an array of the individual Lorenz Curve of each bootstrapped dataset. If \code{return.array = FALSE}, \code{NULL} is returned.
}
\description{
Calculate Lorenz Curve for a vector of data with known frequencies
}
