% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{LRT}
\alias{LRT}
\title{Likelihood Ratio Test (LRT)}
\usage{
LRT(model_candidate, model_ref)
}
\arguments{
\item{model_candidate, model_ref}{objects for which the log likelihood can be extracted via \code{stats::logLik()}. 'model_ref' must have more degrees of freedom than 'model_candidate' for the function to not throw an error. 'model_candidate' should be nested within 'model_ref' for the test to be valid.}
}
\value{
a data.frame of the LRT
}
\description{
Perform likelihood ratio test.
}
