% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{max_scale}
\alias{max_scale}
\title{Maximum Pixel Value of Image}
\usage{
max_scale(object)
}
\arguments{
\item{object}{a cimg, pixset, matrix, or array}
}
\value{
an atomic integer value
}
\description{
Get maximum pixel value of image. Useful for doing arithmetics with image matrix.
}
