% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{melt}
\alias{melt}
\title{Reshape an array to long format}
\usage{
melt(x, drop = FALSE)
}
\arguments{
\item{x}{the array with dimension less than 14.}

\item{drop}{if \code{TRUE}, default is \code{FALSE}, dimensions with length one is removed.}
}
\value{
a data.frame with \code{prod(dim(x))} number of rows and \code{length(dim(x)) + 1} number of columns.
}
\description{
Reshape an array to long format where each row correspond to the index of a value in the nth dimension. The value is stored in the 'val' column. Supports up to a 13 dimension array.
}
