% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_averaging_simulation.R
\name{mid_pts}
\alias{mid_pts}
\alias{mid_pts2}
\title{Find the middle points}
\usage{
mid_pts(x)

mid_pts2(x)
}
\arguments{
\item{x}{a vector of numeric values}
}
\value{
a numeric vector
}
\description{
Find the middle points or the middle points of the middle points of the supplied vector
}
