% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{move_files}
\alias{move_files}
\title{Move files between directories}
\usage{
move_files(from_dir, to_dir, file_names)
}
\arguments{
\item{from_dir}{the source directory}

\item{to_dir}{the destination directory}

\item{file_names}{a character vector of the file names of the files to be moved}
}
\value{
\code{NULL}
}
\description{
Basically a wrapper for \code{file.rename()}
}
