% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{N}
\alias{N}
\title{Get Sample Size}
\usage{
N(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of numeric data}

\item{na.rm}{a logical value indicating whether to drop \code{NA} values}
}
\value{
A numeric value
}
\description{
Basically an alias for \code{length()}
}
