% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nearest_bin}
\alias{nearest_bin}
\title{Find nearest bin}
\usage{
nearest_bin(x, grid, value = TRUE)
}
\arguments{
\item{x}{a vector of numeric values}

\item{grid}{the vector of the bins}

\item{value}{If \code{TRUE}, return the bin value, otherwise return the bin id}
}
\value{
a numeric vector
}
\description{
Find the closest bin of a vector
}
