% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{nn_dist}
\alias{nn_dist}
\title{Compute nearest neighbor distance and perform Clark-Evans test}
\usage{
nn_dist(
  object,
  alternative = c("two.sided", "clustered", "regular", "greater", "lesser"),
  silent = FALSE
)
}
\arguments{
\item{object}{a matrix with two columns (x and y coordinates) or a 'split_herb' object}

\item{alternative}{the alternative hypothesis}

\item{silent}{If \code{TRUE}, the result of the test is not printed}
}
\value{
If \code{silent = TRUE}, the nearest neighbor distances are returned along with the result of the Clark-Evans test. If \code{silent = TRUE}, only the nearest neighbor distances are returned.
}
\description{
Compute nearest neighbor distance and perform Clark-Evans (1954) test
}
\references{
Clark, P. J., and F. C. Evans. 1954. Distance to Nearest Neighbor as a Measure of Spatial Relationships in Populations. Ecology 35:445–453.
}
