% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optim2}
\alias{optim2}
\title{Maximum Likelihood Optimization With Error Handling}
\usage{
optim2(init, fn, method, lower, upper, silent = TRUE, ...)
}
\arguments{
\item{init}{initial values for the parameters to be optimized over}

\item{fn}{function for which to find the minimum}

\item{method}{method allowed by \code{stats::optim()}, as well as \code{nlm} and \code{nlminb}}

\item{lower, upper}{lower and upper bounds}

\item{silent}{if \code{TRUE}, no message is returned. Default to \code{TRUE}.}

\item{...}{extra arguments passed to \code{optim()}, \code{nlm()}, or \code{nlminb()}}
}
\value{
a list
}
\description{
A wrapper function of \code{stats::optim()}, \code{stats::nlm()}, and \code{stats::nlminb()} with error handling. Intended as an interval function.
}
