% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ordered_stat}
\alias{ordered_stat}
\title{Ordered statistics}
\usage{
ordered_stat(x, n)
}
\arguments{
\item{x}{the vector}

\item{n}{the nth order}
}
\value{
an atomic numeric vector
}
\description{
Find the nth ordered statistics
}
