% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overlaps_zero}
\alias{overlaps_zero}
\title{Check if values overlaps zero}
\usage{
overlaps_zero(...)
}
\arguments{
\item{...}{Numeric vectors of any length}
}
\value{
an atomic logical value
}
\description{
Check if values overlaps zero and returns a logical value
}
