% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{pair_dist}
\alias{pair_dist}
\title{Find All Pairwise Distance}
\usage{
pair_dist(object)
}
\arguments{
\item{object}{An object of class 'split_herb', or a n by 2 matrix of centroid position as a matrix, array or data.frame.}
}
\value{
an n choose 2 by 3 matrix of pairwise distance, where n is the number of centroids.
}
\description{
Get all pairwise distance between centroids
}
