% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{plot.cimg}
\alias{plot.cimg}
\title{Plot Image Using Base Graphics Using Missing Value Handling}
\usage{
\method{plot}{cimg}(
  x,
  frame,
  xlim = c(1, width(x)),
  ylim = c(height(x), 1),
  xlab = "x",
  ylab = "y",
  rescale = TRUE,
  colourscale = NULL,
  colorscale = NULL,
  interpolate = TRUE,
  axes = TRUE,
  main = "",
  xaxs = "i",
  yaxs = "i",
  asp = 1,
  col.na = rgb(0, 0, 0, 0),
  ...
)
}
\arguments{
\item{x}{the image}

\item{frame}{which frame to display, if the image has depth > 1}

\item{xlim}{x plot limits (default: 1 to width)}

\item{ylim}{y plot limits (default: 1 to height)}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{rescale}{rescale pixel values so that their range is \link{0,1}}

\item{colourscale, colorscale}{an optional colour scale (default is gray or rgb)}

\item{interpolate}{should the image be plotted with antialiasing (default \code{TRUE})}

\item{axes}{whether to draw axes (default \code{TRUE})}

\item{main}{main title}

\item{xaxs, yaxs}{The style of axis interval calculation to be used for the axes. See \code{?par}}

\item{asp}{aspect ratio.}

\item{col.na}{which colour to use for \code{NA} values, as R rgb code. The default is "rgb(0,0,0,0)", which corresponds to a fully transparent colour.}

\item{...}{additional arguments passed to \code{plot.default()}}
}
\description{
plot image the same way as \code{imager::plot.cimg()}. The only difference is that the \code{rescale} is automatically set to \code{FALSE} when the image contains missing pixel values. see \code{?imager::plot.cimg()} for more details.
}
\note{
The code for the function is obtained from Simon Barthelme's imager:::plot.singleton(). version 0.42.13.
}
