% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{plot.lc}
\alias{plot.lc}
\title{Plot Lorenz Curve Using Base Graphics}
\usage{
\method{plot}{lc}(
  x,
  spaghetti = FALSE,
  main = "Lorenz curve",
  xlab = "Proportion of x",
  ylab = "Cumulative proprotion x",
  lc.lwd = 2,
  lc.col = "red",
  identity.lwd = 3,
  identity.col = "black",
  ...
)
}
\arguments{
\item{x}{an object of class 'lc' generated by \code{lorenz_curve()}}

\item{spaghetti}{if \code{TRUE}, a spaghetti plot of all simulations in \code{boot_array} is plotted. If an integer value is supplied, the number of curves up to the total number of simulations in \code{boot_array} will be plotted. If \code{FALSE} (default), the 95\% confidence interval will be plotted instead.}

\item{main}{the title of the plot}

\item{xlab, ylab}{the x or y axes label}

\item{lc.lwd, identity.lwd}{the thickness of the Lorenz Curve or the identity line (representing perfect equality).}

\item{lc.col, identity.col}{the color of the Lorenz Curve or the identity line (representing perfect equality).}

\item{...}{extra arguments passed to \code{plot.default()}}
}
\value{
\code{NULL}
}
\description{
Plot Lorenz Curve for an object of class 'lc'. If available, uncertainty from the bootstrap simulation is also plotted either as 95\% confidence intervals or as a spaghetti plot.
}
