% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{plot.singleton}
\alias{plot.singleton}
\title{Plot one dimensional image}
\usage{
\method{plot}{singleton}(x, ...)
}
\arguments{
\item{x}{the image}

\item{...}{additional arguments passed to \code{plot.default()}}
}
\description{
plot one dimensional image with pixel value plotted as a line. Internal function of \code{plot.cimg()}.
}
\note{
The code for the function is obtained from Simon Barthelme's \code{imager:::plot.singleton()}. version 0.42.13.
}
