% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{plot.split_herb}
\alias{plot.split_herb}
\title{Display A List of Images Using Base Graphics}
\usage{
\method{plot}{split_herb}(x, prompt = TRUE, ...)
}
\arguments{
\item{x}{a 'split_herb' object}

\item{prompt}{if \code{TRUE} (default), display a prompt for user to select whether to proceed with plotting when the number of holes is above 20.}

\item{...}{extra arguments passed to \code{spatstat.geom::plot.imlist()}}
}
\description{
Plot different regions of herbivory form an object of class 'split_herb'
}
