% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_colors}
\alias{plot_colors}
\title{Plot colors}
\usage{
plot_colors(
  x,
  label = NULL,
  label_size = 3,
  boarder_color = "black",
  background = "grey",
  linear = FALSE
)
}
\arguments{
\item{x}{a vector of hex code or acceptable ggplot color names}

\item{label}{if \code{NULL} (default), the supplied character string via \code{x} is overlayed on the colored tiles. If \code{NA} or \code{FALSE}, nothing is displayed}

\item{label_size}{The size of the label}

\item{boarder_color}{The color of the boarder of the tiles}

\item{background}{The color of the background}

\item{linear}{If \code{TRUE}, the tiles are displayed in one row. Default is \code{FALSe}.}
}
\value{
A ggplot
}
\description{
Display the color or hex code via \code{ggplot2::geom_tile()}
}
