% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Generate Empirical Cumulative Distribution and Histogram Plot}
\usage{
plot_distributions(
  data.list,
  type = c("ecdf", "hist"),
  by = 0.1,
  ecdf.xlim = c(0, 1),
  ...
)
}
\arguments{
\item{data.list}{a list of numeric vectors}

\item{type}{a character vector indicating which plots to display. Acceptable options are "ecdf" and "hist".}

\item{by}{the bin size}

\item{ecdf.xlim}{the x-axis bounds of the ECDF plot.}

\item{...}{additional arguments passed to \code{stats::plot.ecdf()}}
}
\description{
Generate ECDF and histogram plot for each vector in a list
}
