% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pre_cmp_fun}
\alias{pre_cmp_fun}
\title{Pre-compile custom functions}
\usage{
pre_cmp_fun(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
Pre-compile all custom functions in the global environment
}
