% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{predict_list}
\alias{predict_list}
\alias{predict.allo_herb_fit}
\alias{predict.generic_null_fit}
\title{Generate Predictions From Fitted Model}
\usage{
predict_list(
  fit.list,
  n.sim = NULL,
  digits = 2,
  return.obs = TRUE,
  obs.raw = FALSE,
  new.param
)

\method{predict}{allo_herb_fit}(object, ..., n.sim = NULL, digits = 2, new.param)

\method{predict}{generic_null_fit}(object, ..., n.sim = NULL, digits = 2, new.param)
}
\arguments{
\item{fit.list}{A list of "allo_herb_fit" objects or "generic_null_fit"}

\item{n.sim}{A numeric value indicating number of points to draw from the neutral distribution for each "allo_herb_fit" object. If \code{NULL}, the original sample size of the data will be used.}

\item{digits}{A numeric value indicating the number of digits to round data values to}

\item{return.obs}{A logical value indicating whether to return observed data as a list in the out put along with the predicted draws}

\item{obs.raw}{A logical value indicating whether to round the observed data; ignored if \code{return.obs} is set to \code{FALSE}.}

\item{new.param}{A vector of named values used to make new predictions. If a parameter is set to \code{NA}, the parameter value is extracted from the "allo_herb_fit" or "generic_null_fit" object.}

\item{object}{An "allo_herb_fit" or "generic_null_fit" object}

\item{...}{additional arguments}
}
\value{
A list of list of vectors of numeric values.
}
\description{
These are convenience functions that generate predicted distribution from a list of fitted models or a single fitted model object
}
