% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_ML_support.R
\name{print.bite_size_fit}
\alias{print.bite_size_fit}
\title{Print Values}
\usage{
\method{print}{bite_size_fit}(x, ...)
}
\arguments{
\item{x}{An object of class 'bite_size_fit'}

\item{...}{additional arguments}
}
\value{
A matrix array
}
\description{
Prints fitted object and returns an invisible matrix array of the fitted model coefficients
}
