% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_model_ML_support.R
\name{print.generic_null_fit}
\alias{print.generic_null_fit}
\title{Print Values}
\usage{
\method{print}{generic_null_fit}(x, ..., trans = "backtransform", digits = 3)
}
\arguments{
\item{x}{a 'generic_null_fit' object}

\item{...}{additional arguments}

\item{trans}{a function that is used to transform the estimated model coefficients. No transformation is applied when set to \code{FALSE}. If set to "backtransform" (default), then the values are back transformed to the original scale.}

\item{digits}{number of digits to display. Default is 3.}
}
\value{
a matrix
}
\description{
Print 'generic_null_fit' objects
}
