% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{print.split_herb}
\alias{print.split_herb}
\title{Print Value}
\usage{
\method{print}{split_herb}(x, ..., digits = 4)
}
\arguments{
\item{x}{a 'split_herb' object}

\item{...}{additional arguments}

\item{digits}{number of digits to display. Default is four.}
}
\value{
a vector of numeric values
}
\description{
print values from a 'split_herb' object
}
