% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{probe_dist}
\alias{probe_dist}
\title{Generate Summary Statistics For A Vector of Data}
\usage{
probe_dist(
  x,
  probes = c("mean", "var", "cv", "Gini", "max", "min", "median", "Skew", "Kurt",
    "q25", "q75", "N", "lac"),
  na.rm = FALSE
)
}
\arguments{
\item{x}{a vector of numeric data}

\item{probes}{a vector of function names, character string, or functions that can be found with \code{match.fun()}. Default probes are \code{mean()}, \code{var()}, \code{cv()}, \code{Gini()}, \code{max()}, \code{min()}, \code{median()}, \code{Skew()}, \code{Kurt()}, \code{q25()}, \code{q75()}, \code{N()}, and \code{Lasym()}.}

\item{na.rm}{a logical value indicating whether to drop \code{NA} values}
}
\value{
A named vector of numeric values
}
\description{
Get values for a set of statistical probes.
}
\examples{
x <- runif(100,0.1,1)
probe_dist(x)
probe_dist(x, probes = "cv")
probe_dist(x, probes = c("sd", "length","se"))
probe_dist(x, probes = c(function(v){sd(v)/length(v)^0.5}))
}
