% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{probe_distribution_list}
\alias{probe_distribution_list}
\title{Generate Summary Statistics For A List of Vectors of Data}
\usage{
probe_distribution_list(
  data.list,
  probes = c("mean", "var", "cv", "Gini", "max", "min", "median", "Skew", "Kurt",
    "q25", "q75", "N", "lac"),
  trivial.rm = TRUE,
  na.rm = FALSE,
  add.id = FALSE
)
}
\arguments{
\item{data.list}{a list of numeric vectors}

\item{probes}{a vector of function names, character string, or functions that can be found with \code{match.fun()}. Default probes are \code{mean()}, \code{var()}, \code{cv()}, \code{Gini()}, \code{max()}, \code{min()}, \code{median()}, \code{Skew()}, \code{Kurt()}, \code{q25()}, \code{q75()}, \code{N()}, and \code{Lasym()}.}

\item{trivial.rm}{a logical value indicating whether to drop rows where the data are all zeros. Default to \code{TRUE}.}

\item{na.rm}{a logical value indicating whether to drop \code{NA} values}

\item{add.id}{An argument for internal use only. Default is \code{FALSE}. Adds a column called "surveyID" from the data.list name supplied.}
}
\value{
A named vector of numeric values
}
\description{
Get values for a set of statistical probes for each vector in a list
}
