% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{px_size_calc}
\alias{px_size_calc}
\title{Calculate Pixel Size}
\usage{
px_size_calc(object = NULL, px.size)
}
\arguments{
\item{object}{a matrix, array, cimg, or pixset that is optionally passed to measure image size. Default is NULL.}

\item{px.size}{A list of the correct format as would be returned by this function, \code{NA}, \code{NULL}, or a character string. see details.}
}
\value{
If \code{px.size} is \code{NA} or \code{NULL}, \code{NULL} is returned. Otherwise, a list is returned.

\code{size}: the length of the side of a single pixel

\code{unit}: unit of size. Currently, only mm is supported.
}
\description{
Calculate pixel size from known parameters or image
}
\details{
The character string must be in the general format "reference unit:value". Other special characters "=" and "-" are also acceptable to delineate "reference unit" from "value".

The "reference unit" specifies the unit of "value". This can be "dpi", "ppi", "mm_x" or "cm_x". Here, "x" (must be separated by the character "_") is a placeholder for a number.

The "value" specifies the image resolution (for dpi or ppi) or how many pixels ("mm_x" or "cm_x") are represented by the reference unit. "width" or "height" may be supplied in place of a number to indicate that the width or height of the image is to be measured from the supplied object instead.
}
\examples{
# Image resolution is 150 dpi or 150 ppi
px_size_calc(NULL, px.size = "dpi:150")
px_size_calc(NULL, px.size = "ppi-150")

# 1234 pixels is known to be 33 mm long.
px_size_calc(NULL, px.size = "mm_33:1234")


# image width (100 pixels) is 33 cm
px_size_calc(matrix(rnorm(10000),nrow = 100), px.size = "cm_33:width")
px_size_calc(NULL, px.size = "cm_33=100") # same thing

}
