% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{q25}
\alias{q25}
\title{Get 25th Percentile}
\usage{
q25(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of numeric data}

\item{na.rm}{a logical value indicating whether to drop \code{NA} values}
}
\value{
A numeric value
}
\description{
Find the lower quantile of a vector of numeric data
}
