% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{q50}
\alias{q50}
\title{Get Median}
\usage{
q50(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of numeric data}

\item{na.rm}{a logical value indicating whether to drop \code{NA} values}
}
\value{
A numeric value
}
\description{
Find median of a vector of numeric data
}
