% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{q75}
\alias{q75}
\title{Get 75th Percentile}
\usage{
q75(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of numeric data}

\item{na.rm}{a logical value indicating whether to drop \code{NA} values}
}
\value{
A numeric value
}
\description{
Find the upper quantile of a vector of numeric data
}
