% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{r2_delta.brmsfit}
\alias{r2_delta.brmsfit}
\title{Delta R2 of Fitted 'brmsfit' Model Predictor}
\usage{
\method{r2_delta}{brmsfit}(
  object,
  var,
  summary = TRUE,
  robust = FALSE,
  probs = c(0.025, 0.975),
  ndraws = NULL,
  draw_ids = NULL,
  ...
)
}
\arguments{
\item{object}{A 'brmsfit' object}

\item{var}{the name of the variable for which to assess the delta R2}

\item{summary}{if \code{TRUE} (default), the posterior draws are summarized}

\item{robust}{if \code{TRUE} (default is FALSE), the median instead of the mean is returned as the estimate}

\item{probs}{The lower and upper interval of posterior summary}

\item{ndraws}{The number of posterior samples used in the simulation, passed to the \code{'ndraws'} argument of \code{brms:::posterior_epred.brmsfit()}. If \code{NULL} (default), all samples are used.}

\item{draw_ids}{An integer vector specifying the posterior draws to be used. If \code{NULL} (default), all draws are used.}

\item{...}{additional arguments}
}
\value{
a matrix of delta r2 draws or a vector of delta r2 summary.
}
\description{
Calculate the Bayesian delta R2 of a specified predictor variable by finding the change in r2 when the predictor variable is set to zero.
}
