% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{r2_partial}
\alias{r2_partial}
\title{Partial R2 of Fitted Model Predictor}
\usage{
r2_partial(object, var, ...)
}
\arguments{
\item{object}{a supported model fit object}

\item{var}{the name of the variable for which to assess the partial R2}

\item{...}{additional parameters}
}
\value{
the partial r2
}
\description{
Generic method of finding the partial R2 of a specified predictor variable by finding the change in r2 when the predictor variable is set to zero.
}
