% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{r2_partial_all.brmsfit}
\alias{r2_partial_all.brmsfit}
\title{Partial R2 of All Fitted 'brmsfit' Model Predictors}
\usage{
\method{r2_partial_all}{brmsfit}(
  object,
  summary = TRUE,
  robust = FALSE,
  probs = c(0.025, 0.975),
  ndraws = NULL,
  draw_ids = NULL,
  ...
)
}
\arguments{
\item{object}{A 'brmsfit' object}

\item{summary}{if \code{TRUE} (default), the posterior draws are summarized}

\item{robust}{if \code{TRUE} (default is FALSE), the median instead of the mean is returned as the estimate}

\item{probs}{The lower and upper interval of posterior summary}

\item{ndraws}{The number of posterior samples used in the simulation, passed to the \code{'ndraws'} argument of \code{brms:::posterior_epred.brmsfit()}. If \code{NULL} (default), all samples are used.}

\item{draw_ids}{An integer vector specifying the posterior draws to be used. If \code{NULL} (default), all draws are used.}

\item{...}{additional arguments passed to \code{r2_partial()}}
}
\value{
a data.frame of partial R2 values
}
\description{
Calculate the Bayesian partial R2 of all predictor variables by finding the change in r2 when the predictor variable is set to zero.
}
