% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{r2_partial_all}
\alias{r2_partial_all}
\title{Partial R2 of All Fitted Model Predictors}
\usage{
r2_partial_all(object, ...)
}
\arguments{
\item{object}{a supported model fit object}

\item{...}{additional parameters passed to \code{r2_partial()}}
}
\value{
a data.frame of partial R2 values
}
\description{
Generic method of finding the partial R2 of all predictor variables by finding the change in r2 when the predictor variable is set to zero.
}
