% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{randomize_leaves}
\alias{randomize_leaves}
\title{Randomize Leaf Herbivory Observations}
\usage{
randomize_leaves(data_list, meta_data, nboot = 1, summarise_plant = TRUE)
}
\arguments{
\item{data_list}{a named list of data to be randomized. The name should match the data set ID provided with the meta_data argument.}

\item{meta_data}{a data.frame of meta data to be used in the randomization procedure. First column of meta_data is the data set ID. Second column of meta_data is the plant ID. Third column of the meta_data is the number of leaves on that plant that has herbivory data.}

\item{nboot}{the number of times to repeat the randomization procedure. Default is one.}

\item{summarise_plant}{if \code{TRUE} (default), the randomized leaf herbivory will be averaged by plant.}
}
\value{
a list of list of numeric vectors
}
\description{
Randomize leaf herbivory observations using provided meta data. Used to generate null distribution where leaves on individual plants are random draws from all leaves within a population or survey.
}
