% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{rep_data.frame}
\alias{rep_data.frame}
\title{Duplicate data.frame n times}
\usage{
rep_data.frame(x, n)
}
\arguments{
\item{x}{the data.frame}

\item{n}{an interger}
}
\value{
a data.frame
}
\description{
Duplicate the rows of a data.frame n times
}
