% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{rotate_90}
\alias{rotate_90}
\title{Rotate array by 90 degrees}
\usage{
rotate_90(img)
}
\arguments{
\item{img}{the image}
}
\value{
an object of the same class
}
\description{
Rotate an array, cimg, or pixset by 90 degrees.
}
