% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{se}
\alias{se}
\title{Calculate Standard Error of The Mean (SE)}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of numeric data}

\item{na.rm}{a logical value indicating whether to drop \code{NA} values}
}
\value{
A numeric value
}
\description{
calculate standard error for a vector of data
}
\details{
The standard error of the mean \eqn{SE}, is calculated as
\deqn{SE(X) = \sqrt{\frac{\matbb{Var}[X]}{n}}}
}
