% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_interval}
\alias{seq_interval}
\title{Evenly spaced sequences between the range of a vector}
\usage{
seq_interval(x, length.out = 300)
}
\arguments{
\item{x}{the vector}

\item{length.out}{the length of output vector}
}
\value{
a numeric vector
}
\description{
Generate evenly spaced sequences between the range of a vector \code{x} using \code{seq()}
}
