% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{slice_eval}
\alias{slice_eval}
\title{Evaluate Function For Each Height, Width, And Depth Element}
\usage{
slice_eval(..., FUN)
}
\arguments{
\item{...}{a set of array, cimg, or pixset}

\item{FUN}{a function to be evaluated, supported by \code{match.fun()}.}
}
\value{
a cimg object
}
\description{
Evaluate a function for each height, width, and depth element share among a set of provided objects or color spectrum arrays.
}
\examples{
img <- image_example() \%>\% thin(3)

# Find the maximum RGB value
out <- slice_eval(G(img),R(img),B(img),FUN = "max")
out
# same thing
all.equal(
  out,
  slice_eval(img,FUN = "max"))

}
