% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{clarkevans}
\alias{clarkevans}
\alias{clarkevans.test}
\alias{hopskel}
\alias{allstats}
\alias{Lest}
\alias{Kest}
\alias{Gest}
\alias{Fest}
\alias{Jest}
\alias{envelope.ppp}
\title{Exported functions from 'spatstat'}
\usage{
clarkevans(...)

clarkevans.test(...)

hopskel(...)

allstats(...)

Lest(...)

Kest(...)

Gest(...)

Fest(...)

Jest(...)

envelope.ppp(...)
}
\arguments{
\item{...}{see the documentation of the \code{spatstat} packages.}
}
\value{
see the documentation of the \code{spatstat} packages.
}
\description{
Essentially a wrapper for functions in the \code{spatstat} packages.
}
