% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{split2mat}
\alias{split2mat}
\title{Convert split_herb Into Matrix}
\usage{
split2mat(split_herb)
}
\arguments{
\item{split_herb}{an object of class 'split_herb'}
}
\value{
a matrix
}
\description{
converts split_herb into matrix
}
