% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{split2ppp}
\alias{split2ppp}
\title{Convert split_herb To Point Pattern}
\usage{
split2ppp(split_herb, pt.type = c("centroid", "px"))
}
\arguments{
\item{split_herb}{an object of class 'split_herb'.}

\item{pt.type}{An atomic character indicating what to turn into a point. Acceptable values are "centroid" for centroids, and "px" for pixels.}
}
\value{
an object of class 'ppp' with pixels as the unit
}
\description{
converts a 'split_herb' object to a 'ppp' object. When \code{pt.type = "centroid"}, centroids that are rejected by \code{ppp()} because they are out of bound are nudged to the nearest within bound location. Usually, these points are on or close to the edge already.
}
