% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{survival_plot}
\alias{survival_plot}
\title{Survival Plot}
\usage{
survival_plot(x, add = FALSE, ...)
}
\arguments{
\item{x}{a vector of numeric values greater than 0}

\item{add}{a logical value indicating whether to overlay points upon an existing plot}

\item{...}{additional arguments passed to \code{points()} or \code{plot()}}
}
\value{
a plot
}
\description{
Generate a log-log survival plot of positive only sample. Can be useful for visualizing the tail of the distribution.
}
\examples{
x <- rlnorm(10000,1,1)
survival_plot(x)

x2 <- rlnorm(10000,1,1.1)
survival_plot(x2, add = TRUE, col = "green", pch = 19)

}
\references{
Aban, I. B., M. M. Meerschaert, and A. K. Panorska. 2006. Parameter Estimation for the Truncated Pareto Distribution. Journal of the American Statistical Association 101:270–277.
}
