% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{sweep_images}
\alias{sweep_images}
\title{Delete image files from a directory}
\usage{
sweep_images(
  dir,
  ask = TRUE,
  recursive = FALSE,
  file_types = c("jpg", "jpeg", "png", "tif")
)
}
\arguments{
\item{dir}{the directory for which image files are searched}

\item{ask}{if \code{TRUE} (default), user is prompted with a double check menu}

\item{recursive}{if \code{TRUE}, all sub-directories of the supplied directories will also be searched}

\item{file_types}{a vector of the file types for which the function finds and deletes. Defaults are \code{c("jpg","jpeg","png","tif")}}
}
\value{
\code{NULL}
}
\description{
Basically a wrapper for \code{file.remove()}
}
