% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{threshold2}
\alias{threshold2}
\title{Threshold Image With Missing Values And Error Handling}
\usage{
threshold2(
  im,
  thr = "auto",
  approx = TRUE,
  adjust = 1,
  approx.res = 10000,
  km = c(1L, 2L),
  return.thr.only = FALSE,
  thr.exact = FALSE
)
}
\arguments{
\item{im}{the image}

\item{thr}{a threshold, either numeric, or "auto", or string for quantities. If \code{thr.exact} is \code{TRUE}, the value must be numeric.}

\item{approx}{skip pixels when computing quantiles in large images (defaults to \code{TRUE})}

\item{adjust}{use to adjust the automatic threshold: if the auto-threshold is at k, effective threshold will be at adjust*k (default 1)}

\item{approx.res}{the number of pixels used in threshold approximation. Default is 10000.}

\item{km}{an integer vector of length two. The first value selects the the boundary rank in increasing means identified by kmeans clustering (defaults to 1, selecting the boundary that separates the two clusters with the lowest mean). The second value selects the k number of clusters to form (defaults to 2). Ignored when not auto-thresholding.}

\item{return.thr.only}{if \code{TRUE} (default to \code{FALSE}), return the calculated numeric threshold only, otherwise a selected pixset of the image will be returned.}

\item{thr.exact}{if \code{TRUE}, the exact threshold supplied by \code{thr} is used. Used internally to reduce redundant computation.}
}
\value{
a pixet with select pixels if \code{return.thr.only = FLASE}, otherwise, a numeric value is returned
}
\description{
Basically the same thing as \code{imager::threshold()}, but handles missing values in the cimg object, preserves "px.size" attribute, increases the approximation resolution, and handles error generated by k-means approximation.
}
\details{
see \code{?imager::threshold()}
}
\note{
The code for much of the function is obtained from Simon Barthelme's \code{imager::threshold()}. version 0.42.13.
}
