% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weirdo_distributions.R
\name{dtpareto}
\alias{dtpareto}
\alias{rtpareto}
\title{Truncated Pareto Distribution}
\usage{
dtpareto(x, a, b, endpoint = 1, log = FALSE)

rtpareto(n, a, b, endpoint = 1)
}
\arguments{
\item{x}{a vector of quantities}

\item{a}{the exponent of the Pareto distribution}

\item{b}{the minimum value of \code{x}}

\item{endpoint}{a numeric value where the Pareto distribution is truncated. Default value is 1.}

\item{log}{a logical value indicating whether to log transform the probabilities. Default is \code{FALSE}.}

\item{n}{integer value indicating the number of observations to generate}
}
\value{
a vector of numeric values
}
\description{
Density function and random number generation of a three parameter truncated Pareto distribution
}
\details{
The density function is defined as:

If \eqn{x < B}
\deqn{P(x) = \frac{a b^a}{x^{a+1}}\frac{1}{F(x = B; a, b)},}
otherwise,
\deqn{P(x) = 0,}
where \eqn{B} is the end point where the Pareto distribution is truncated, \eqn{b} is the minimum value of x, \eqn{a} is the exponent of the Pareto distribution, and \eqn{F(x = B; a, b)} is the CDF of the Pareto distribution defined as
\deqn{F(x) = 1 - (\frac{b}{x})^a}
}
