% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unscale}
\alias{unscale}
\title{Unscale variable}
\usage{
unscale(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a function
}
\description{
Generate a function from a vector that can back transform a Z-score transformed vector to its original scale.
}
