% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{validate_file_type}
\alias{validate_file_type}
\title{Validate file extension}
\usage{
validate_file_type(x, check_if = c("jpg", "jpeg"), rm = FALSE, stop = TRUE)
}
\arguments{
\item{x}{a vector of file names}

\item{check_if}{a vector of acceptable file types. Defaults are \code{c("jpg", "jpeg")}}

\item{rm}{If \code{TRUE} (default is \code{FALSE}), remove file names that do not match the acceptable file extensions.}

\item{stop}{If \code{TRUE} (default), throw an error when unexpected file types are encountered. Otherwise, a warning is returned.}
}
\value{
a vector of characters
}
\description{
Check if the supplied file names have the correct file extensions
}
