% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{watershed2}
\alias{watershed2}
\title{Compute Watershed Transform With Gray Scale and Color Image Support}
\usage{
watershed2(seed, p, spectrum, thr = "auto", approx.res = 10000, km = c(1L, 2L))
}
\arguments{
\item{seed}{an image}

\item{p}{priority map}

\item{spectrum}{number of spectrum.}

\item{thr}{a threshold, either numeric, or "auto", or string for quantities.}

\item{approx.res}{the number of pixels used in threshold approximation. Default is 10000.}

\item{km}{an integer vector of length two. The first value selects the the boundary rank in increasing means identified by kmeans clustering (defaults to 1, selecting the boundary that separates the two clusters with the lowest mean). The second value selects the k number of clusters to form (defaults to 2). Ignored when auto thresholding.}
}
\value{
a pixset
}
\description{
Internal function of \code{clean_leaf()}. A wrapper of \code{watershed()} and \code{threshold()} that handles both gray scale and color images. If available, the blue spectrum of the transformed image is used for thresholding, as blue is a rare color among plats.
}
