% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{var_wt}
\alias{var_wt}
\alias{mean_wt}
\title{Weighted central moments}
\usage{
var_wt(x, w, biased = FALSE)

mean_wt(x, w)
}
\arguments{
\item{x}{a vector of numeric values}

\item{w}{weights of the same length as x}

\item{biased}{if \code{TRUE}, returns biased estimate}
}
\value{
an atomic numeric value
}
\description{
Computes weighed central moments
}
