% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_null_functions.R
\name{dzoibeta}
\alias{dzoibeta}
\alias{rzoibeta}
\title{Zero-One-Inflated Beta Distribution}
\usage{
dzoibeta(x, p0, p1, alpha, beta, log = FALSE)

rzoibeta(n, p0, p1, alpha, beta)
}
\arguments{
\item{x}{A vector of numeric values}

\item{p0}{The probability that \eqn{x = 0}. Must be between 0 and 1}

\item{p1}{The conditional probability that \eqn{x = 1} given \eqn{x \neq 0}. Must be between 0 and 1.}

\item{alpha, beta}{The shape parameters of the beta distribution. Must be greater than 0.}

\item{log}{a logical value indicating whether to log transform the probabilities. Default is \code{FALSE}.}

\item{n}{number of observations to generate}
}
\value{
A vector of probability densities
}
\description{
Density function of a four parameter zero-one-inflated beta distribution. The beta distribution does not have defined probabilities at 0 or 1 so this extension takes care of that.
}
\details{
The density function is defined as:

If \eqn{x = 0}
\deqn{f(x) = p_0}
If \eqn{x = 1}
\deqn{f(x) = p_1 (1 - p_0)}
If \eqn{x \neq 0,1}
\deqn{f(x) = (1 - p_1) (1 - p_0) \frac{x^{\alpha-1}(1-x)^{\beta-1}}{B(\alpha, \beta)}}
}
