% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_quote}
\alias{add_quote}
\title{Add quotes around a character string}
\usage{
add_quote(x)
}
\arguments{
\item{x}{a vector}
}
\value{
an atomic character vector
}
\description{
Add quotes around a character string
}
