% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{adjust_prop}
\alias{adjust_prop}
\title{Adjust Proportion Data}
\usage{
adjust_prop(
  x,
  nudge.method = c("replace", "none", "smithson", "add", "subtract", "drop"),
  nudge.size = c("macmillan", "warton_min", "warton_max"),
  trans = c("identity", "logit", "empirical_logit", "probit", "log", "sqarcsin"),
  bounds = c("both", "zero", "one"),
  na.action = c("ignore", "remove", "as.is", "fail")
)
}
\arguments{
\item{x}{a vector of numeric data bounded between 0 and 1 inclusive.}

\item{nudge.method}{a character value indicating the method of getting rid of 0's and 1's. Valid options are "replace" (default), "none", "smithson", "add", "subtract", and "drop". see details.}

\item{nudge.size}{a numeric value of the size of a small nudge (\eqn{\epsilon}). Valid character values can also be supplied to choose the method for estimating \eqn{\epsilon}. Valid methods are "macmillan" (default), "warton_min", and "warton_max". see details.}

\item{trans}{a character value indicating the transformation to be applied to the nudged data. Valid options are "identity" (default-- no transformation), "logit", "empirical_logit", "probit", "log", and "sqarcsin". See details.}

\item{bounds}{a character value indicating bounds for which the \code{nudge.method} is applied to. Valid options are "both" (default), "zero", and "one". Only relevant for \code{nudge.method} set as "replace" or "drop".}

\item{na.action}{a character value indicating how to handle missing values. Valid options are "ignore", "remove", "as.is", and "fail". See details.}
}
\value{
a vector of numeric value with the same length as \code{x}
}
\description{
Adjusts proportion data to get rid of 0's and 1's, and optionally transforms the resulting adjusted proportion.
}
\details{
\strong{Add commentary}
\subsection{Handling 1's and 0's.}{
\subsection{Nudge method}{

\code{replace}: Replace 0 by \eqn{\epsilon} and 1 by \eqn{1-\epsilon}

\code{none}: No method applied. Can be used along with the empirical logit transformation.

\code{smithson}: the vector is transformed according to the recommendation of Smithson & Verkuilen (2006): \eqn{\frac{x(n-1) + \epsilon}{n}}

\code{add}: the vector is transformed as \eqn{x + \epsilon}

\code{subtract}: the vector is transformed as \eqn{x - \epsilon}

\code{drop}: 0's and/or 1's are dropped from the vector
}

\subsection{Epsilon}{

\code{macmillan}: estimated as \eqn{\frac{1}{2n}} per the recommendation of Macmillan & Creelman (2004), where \eqn{n} is the sample size.

\code{warton_min}: estimated as the smallest non-zero value in the vector x per Warton & Hui (2011).
\code{warton_max}: estimated as one minus the largest non-one value in the vector x, useful when \code{warton_min} is too big (Warton & Hui 2011).
}

}

\subsection{Transformation}{

\code{identity}: no transformation

\code{logit}: \eqn{\ln{\frac{x}{1-x}}}. Undefined when \eqn{x = 1} or \eqn{x = 0}. Recommended by Warton & Hui (2011) over the square root arcsine transformation when transforming true proportions for analysis that assumes a Gaussian distribution.

\code{empirical_logit}: \eqn{\ln{\frac{x+\epsilon}{1-x+\epsilon}}}. A modified logit transformation recommend by Warton & Hui (2011) when 0's or 1's are present in the data.

\code{probit}: \eqn{\int_{-\infty}^{x}\frac{1}{\sqrt{2 \pi}}\exp{-\frac{t^2}{2}}dt}

\code{log}: \eqn{\ln(x)}

\code{sqarcsin}: \eqn{\arcsin(\sqrt{x})}
}

\subsection{Missing Value Handling}{

\code{ignore}: Function will ignore missing values in the data and return a vector of adjusted values with the original missing values in place.

\code{remove}: Function will remove missing values in the data and return a shorter vector or adjusted values.

\code{as.is}: Function will perform the computation with the missing values, possibly returning a vector of \code{NA}s.

\code{fail}: Function will throw an error if the data contains any missing value.
}
}
\examples{
x <- c(0,runif(0,0,1),NA,1,1,1)
adjust_prop(x, nudge.method = "replace", nudge.size = 0.01)

adjust_prop(x, nudge.method = "replace", nudge.size = "macmillan", trans = "logit")


}
\references{
Kubinec, R. 2022. Ordered Beta Regression: A Parsimonious, Well-Fitting Model for Continuous Data with Lower and Upper Bounds. Political Analysis:1–18.

Macmillan, N. A., and C. D. Creelman. 2004. Detection Theory: A User’s Guide. Second edition. Psychology Press, New York.

Smithson, M., and J. Verkuilen. 2006. A better lemon squeezer? Maximum-likelihood regression with beta-distributed dependent variables. Psychological Methods 11:54–71.

Warton, D. I., and F. K. C. Hui. 2011. The arcsine is asinine: the analysis of proportions in ecology. Ecology 92:3–10.
}
