% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neutral_model_ML_support.R
\name{AICc}
\alias{AICc}
\title{Corrected Akaike's An Information Criterion (AICc)}
\usage{
AICc(object = NULL, logLik = NULL, k = 2)
}
\arguments{
\item{object}{The object from which log likelihood can be extracted}

\item{logLik}{An object of calss 'logLik'. Ignored if an object is supplied.}

\item{k}{the penalty on model complexity. Default is 2.}
}
\value{
A numeric value
}
\description{
Calculate AICc taking into account small sample size bias
}
\details{
The formulat for AICc is given by:
\deqn{AIC_c = -2\ln{\mathcal{L}(x)} + k  d \frac{n}{n - d - 1}}
\eqn{d} is the number of parameters, \eqn{k} is the penalty, \eqn{n} is the sample size, and \eqn{\mathcal{L}(x)} is the likelihood.
}
