% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neutral_model_ML_support.R
\name{coef.allo_herb_fit}
\alias{coef.allo_herb_fit}
\title{Extract Model Coefficients}
\usage{
\method{coef}{allo_herb_fit}(object, ..., backtransform = TRUE)
}
\arguments{
\item{object}{An 'allo_herb_fit' object}

\item{...}{additional arguments}

\item{backtransform}{A logic value indicating whether to back transform coefficients from the scale the coefficient was estimated at to the scale the coefficient is parameterized as for the neutral model. Default is TRUE.}
}
\value{
A named vector
}
\description{
extract model coefficients from allo_herb_fit objects
}
