% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{crop_leaf}
\alias{crop_leaf}
\title{Crop Out Non-Leaf Pixels}
\usage{
crop_leaf(object, thr = "kmeans", invalid = NA, rgb.index = 3L, px.size = NA)
}
\arguments{
\item{object}{a cimg, pixset, or matrix, or array to be cropped.}

\item{thr}{a threshold, either numeric, "kmeans", or "otsu", or string for quantiles. If \code{FALSE} (default is "kmeans"), no thesholding will be performed prior to cropping. Useful for when the colors are slightly impure.}

\item{invalid}{the value to replace non-leaf pixels as. Default is \code{NA}.}

\item{rgb.index}{an integer value \code{1L}, \code{2L}, or \code{3L}, indicating red, green, or blue colors respectively to be identified as non-leaf pixels.}

\item{px.size}{value passed to \code{px_size_calc()}. When set to \code{NA} (default), the 'px.size' attribute is extracted from the supplied \code{object}.}
}
\value{
a cropped 'cimg' object with only one color spectrum
}
\description{
Crop out non-leaf pixels marked by the color blue as \code{NA}s. Non-leaf pixels need to be cropped out for further analyses.
}
\examples{
file_path <- system.file("extdata/mock_leaf8.png",package="herbivar")
img <- load.image(file_path) \%>\% add_px_size("dpi:300")
plot(img)
cropped_img <- crop_leaf(img)
plot(cropped_img)

analyze_holes(cropped_img) \%>\% print()

}
