% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{invert}
\alias{invert}
\title{Invert colors}
\usage{
invert(object)
}
\arguments{
\item{object}{an image, pixset, or imlist}
}
\value{
the same object with inverted colors
}
\description{
Invert colors of an image. Works with any object type that is supported by \code{imager::imeval()}.
}
