% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{LRT}
\alias{LRT}
\title{Likelihood Ratio Test (LRT)}
\usage{
LRT(model_candidate, model_null)
}
\arguments{
\item{model_candidate, model_null}{objects for which the log likelihood can be extracted via \code{stats::logLik()}. Model_null must have more degrees of freedom than model_candidate.}
}
\value{
a data.frame of the LRT
}
\description{
Perform likelihood ratio test
}
