% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herb_image_helpers.R
\name{mat2ppp}
\alias{mat2ppp}
\title{Convert Matrix Into Point Pattern}
\usage{
mat2ppp(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
a 'ppp' object with pixels as the unit.
}
\description{
convert matrix into point pattern. Non-leaf parts, remaining leaf parts, and eaten leaf parts of the image matrix must be indicated by \code{NA}s, 0, and any number greater than 0 respectively.
}
