% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optim2}
\alias{optim2}
\title{General-Purpose Optimization With Error Handling}
\usage{
optim2(silent = TRUE, ...)
}
\arguments{
\item{silent}{if \code{TRUE}, no message is returned. Default to \code{TRUE}.}

\item{...}{extra arguments passed to \code{optim()}}
}
\value{
a list
}
\description{
A wrapper function of \code{stats::optim()} with error handling.
}
