% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{plot.imlist}
\alias{plot.imlist}
\title{Plot image list}
\usage{
\method{plot}{imlist}(x, main = "", main.panel = NULL, interpolate = FALSE, ...)
}
\arguments{
\item{x}{an object of class 'imlist'}

\item{main}{a character string for the plot title}

\item{main.panel}{a vector of character string for the panel title}

\item{interpolate}{a logical value indicating whether to interpolate the raster}

\item{...}{additional arguments passed to \code{spatstat.geom::plot.imlist()}}
}
\value{
\code{NULL}
}
\description{
plot image list via \code{spatstat.geom::plot.imlist()}
}
