% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate_distributions.R
\name{plot_distributions}
\alias{plot_distributions}
\title{Generate Empirical Cumulative Distribution and Histogram Plot}
\usage{
plot_distributions(
  data_list,
  type = c("ecdf", "hist"),
  by = 0.1,
  plot_smooth = TRUE,
  xlim = c(0, 1),
  return_transformed = FALSE,
  ...
)
}
\arguments{
\item{data_list}{a list of numeric vectors}

\item{type}{a character vector indicating which plots to display. Acceptable options are "ecdf" and "hist".}

\item{by}{the bin size}

\item{plot_smooth}{if \code{TRUE} (default), display smoothed line(s) for zoomed-in histogram on the log scale.}

\item{xlim}{the x-axis bounds.}

\item{return_transformed}{if \code{TRUE}, return a zoomed-in histogram on the log scale.}

\item{...}{additional arguments passed to \code{stats::plot.ecdf()}}
}
\description{
Generate ECDF and histogram plot for each vector in a list. If the range specified by \code{xlim} does not contain all the values in the data vector, an error will be returned.
}
