% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral_model_ML_support.R
\name{print.allo_herb_fit}
\alias{print.allo_herb_fit}
\title{Print Values}
\usage{
\method{print}{allo_herb_fit}(x, ..., trans = "backtransform", digits = 3)
}
\arguments{
\item{x}{An object of class 'allo_herb_fit'}

\item{...}{additional arguments}

\item{trans}{a function that is used to transform the estimated model coefficients. No transformation is applied when set to \code{FALSE}. If set to "backtransform" (default), then the values are back transformed to the original scale.}

\item{digits}{A numeric value indicating the number of digits to be displayed in the model coefficients. Default is 3.}
}
\value{
A matrix array
}
\description{
Prints fitted object and returns an invisible matrix array of the fitted model coefficients
}
