% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_herb_image_support.R
\name{thin}
\alias{thin}
\title{Reduce Pixel Density}
\usage{
thin(object, thin = 3)
}
\arguments{
\item{object}{an object of class 'cimg', 'pixset', 'matrix', or 'array'}

\item{thin}{a numeric value indicating the square root of the factor by which to reduce pixel density. Default is 3, which reduces pixel density by 9 times.}
}
\value{
an object of the same class
}
\description{
reduce pixel density
}
